### 2023-08-30 #5790 @danielskatz
👋 @rfj82982 - thanks for your submission.  I'll now look for an editor.

While I do so, please fix the incorrect DOIs by removing the prefix. Please feel free to make changes to your .bib file, then use the command `@editorialbot check references` to check again, and the command `@editorialbot generate pdf` when the references are right to make a new PDF. editorialbot commands need to be the first entry in a new comment.
