+++ submissionjoss/paper.bib
@@ -26,13 +26,14 @@
 	title = {Distributed-Memory Simulations of Turbulent Flows on Modern GPU Systems Using an Adaptive Pencil Decomposition Library},
 	url = {https://doi.org/10.1145/3539781.3539797},
 	year = {2022},
-	bdsk-url-1 = {https://doi.org/10.1145/3539781.3539797}}
+	bdsk-url-1 = {https://doi.org/10.1145/3539781.3539797}
+}
 
 @article{ZHU2018199,
 	author = {Xiaojue Zhu and Everett Phillips and Vamsi Spandan and John Donners and Gregory Ruetsch and Joshua Romero and Rodolfo Ostilla-M{\'o}nico and Yantao Yang and Detlef Lohse and Roberto Verzicco and Massimiliano Fatica and Richard J.A.M. Stevens},
 	date-added = {2022-09-27 03:54:12 +0100},
 	date-modified = {2022-09-27 03:54:12 +0100},
-	doi = {https://doi.org/10.1016/j.cpc.2018.03.026},
+	doi = {10.1016/j.cpc.2018.03.026},
 	issn = {0010-4655},
 	journal = {Computer Physics Communications},
 	keywords = {GPU, Parallelization, Turbulent flow, Finite-difference scheme, Rayleigh--B{\'e}nard convection, Plane Couette flow},
@@ -42,13 +43,13 @@
 	volume = {229},
 	year = {2018},
 	bdsk-url-1 = {https://www.sciencedirect.com/science/article/pii/S0010465518300985},
-	bdsk-url-2 = {https://doi.org/10.1016/j.cpc.2018.03.026}}
+	bdsk-url-2 = {10.1016/j.cpc.2018.03.026}}
 
 @article{BARTHOLOMEW2020100550,
 	author = {Paul Bartholomew and Georgios Deskos and Ricardo A.S. Frantz and Felipe N. Schuch and Eric Lamballais and Sylvain Laizet},
 	date-added = {2022-09-27 03:52:38 +0100},
 	date-modified = {2022-09-27 03:52:38 +0100},
-	doi = {https://doi.org/10.1016/j.softx.2020.100550},
+	doi = {10.1016/j.softx.2020.100550},
 	issn = {2352-7110},
 	journal = {SoftwareX},
 	keywords = {High-order finite-difference, Turbulence, Cartesian mesh, High Performance Computing, Computational Fluid Dynamics},
@@ -58,7 +59,7 @@
 	volume = {12},
 	year = {2020},
 	bdsk-url-1 = {https://www.sciencedirect.com/science/article/pii/S2352711019303620},
-	bdsk-url-2 = {https://doi.org/10.1016/j.softx.2020.100550}}
+}
 
 @inproceedings{li_2010_2decomp,
 	author = {Li, Ning and Laizet, Sylvain},
@@ -66,47 +67,58 @@
 	date-added = {2022-09-27 03:48:20 +0100},
 	date-modified = {2022-09-27 03:48:49 +0100},
 	pages = {1--13},
-	title = {2decomp \& fft-a highly scalable 2d decomposition library and fft interface},
-	year = {2010}}
+        title={2DECOMP\&FFT - A Highly Scalable 2D Decomposition Library and FFT Interface},
+	year = {2010},
+}
 
 @article{godoy2020adios,
-  title={Adios 2: The adaptable input output system. a framework for high-performance data management},
-  author={Godoy, William F and Podhorszki, Norbert and Wang, Ruonan and Atkins, Chuck and Eisenhauer, Greg and Gu, Junmin and Davis, Philip and Choi, Jong and Germaschewski, Kai and Huck, Kevin and others},
-  journal={SoftwareX},
-  volume={12},
-  pages={100561},
-  year={2020},
[…]
