### 2023-09-01 c54ddfa
Profile transpose operations when no MPI communication is performed
