+++ src/transpose_x_to_y.f90
@@ -24,6 +24,10 @@
      integer :: istat, nsize
 #endif
 
+#ifdef PROFILER
+     if (decomp_profiler_transpose) call decomp_profiler_start("transp_x_y_r")
+#endif
+
      if (dims(1) == 1) then
 #if defined(_GPU)
         nsize = product(decomp%xsz)
@@ -38,6 +42,10 @@
         call transpose_x_to_y_real(src, dst, decomp)
      end if
 
+#ifdef PROFILER
+     if (decomp_profiler_transpose) call decomp_profiler_end("transp_x_y_r")
+#endif
+
   end subroutine transpose_x_to_y_real_long
 
   subroutine transpose_x_to_y_real(src, dst, decomp)
@@ -50,10 +58,6 @@
 
      integer :: s1, s2, s3, d1, d2, d3
      integer :: ierror
-
-#ifdef PROFILER
-     if (decomp_profiler_transpose) call decomp_profiler_start("transp_x_y_r")
-#endif
 
      s1 = SIZE(src, 1)
      s2 = SIZE(src, 2)
@@ -113,11 +117,6 @@
                             decomp%y1dist, decomp)
 #endif
 
-#ifdef PROFILER
-     if (decomp_profiler_transpose) call decomp_profiler_end("transp_x_y_r")
-#endif
-
-     return
   end subroutine transpose_x_to_y_real
 
   subroutine transpose_x_to_y_complex_short(src, dst)
@@ -141,6 +140,11 @@
 #if defined(_GPU)
      integer :: istat, nsize
 #endif
+
+#ifdef PROFILER
+     if (decomp_profiler_transpose) call decomp_profiler_start("transp_x_y_c")
+#endif
+
      if (dims(1) == 1) then
 #if defined(_GPU)
         nsize = product(decomp%xsz)
@@ -155,6 +159,10 @@
         call transpose_x_to_y_complex(src, dst, decomp)
      end if
 
+#ifdef PROFILER
+     if (decomp_profiler_transpose) call decomp_profiler_end("transp_x_y_c")
+#endif
+
   end subroutine transpose_x_to_y_complex_long
 
   subroutine transpose_x_to_y_complex(src, dst, decomp)
@@ -167,10 +175,6 @@
 
      integer :: s1, s2, s3, d1, d2, d3
      integer :: ierror
-
-#ifdef PROFILER
-     if (decomp_profiler_transpose) call decomp_profiler_start("transp_x_y_c")
-#endif
 
      s1 = SIZE(src, 1)
      s2 = SIZE(src, 2)
@@ -231,11 +235,6 @@
                                decomp%y1dist, decomp)
 #endif
 
-#ifdef PROFILER
-     if (decomp_profiler_transpose) call decomp_profiler_end("transp_x_y_c")
-#endif
-
-     return
   end subroutine transpose_x_to_y_complex
 
   subroutine mem_split_xy_real(in, n1, n2, n3, out, iproc, dist, decomp)
