+++ src/transpose_y_to_z.f90
@@ -24,6 +24,10 @@
      integer :: istat, nsize
 #endif
 
+#ifdef PROFILER
+     if (decomp_profiler_transpose) call decomp_profiler_start("transp_y_z_r")
+#endif
+
      if (dims(2) == 1) then
 #if defined(_GPU)
         nsize = product(decomp%ysz)
@@ -37,6 +41,10 @@
      else
         call transpose_y_to_z_real(src, dst, decomp)
      end if
+
+#ifdef PROFILER
+     if (decomp_profiler_transpose) call decomp_profiler_end("transp_y_z_r")
+#endif
 
   end subroutine transpose_y_to_z_real_long
 
@@ -54,10 +62,6 @@
 
      integer :: s1, s2, s3, d1, d2, d3
      integer :: ierror
-
-#ifdef PROFILER
-     if (decomp_profiler_transpose) call decomp_profiler_start("transp_y_z_r")
-#endif
 
      s1 = SIZE(src, 1)
      s2 = SIZE(src, 2)
@@ -134,11 +138,6 @@
 
 #endif
 
-#ifdef PROFILER
-     if (decomp_profiler_transpose) call decomp_profiler_end("transp_y_z_r")
-#endif
-
-     return
   end subroutine transpose_y_to_z_real
 
   subroutine transpose_y_to_z_complex_short(src, dst)
@@ -163,6 +162,10 @@
      integer :: istat, nsize
 #endif
 
+#ifdef PROFILER
+     if (decomp_profiler_transpose) call decomp_profiler_start("transp_y_z_c")
+#endif
+
      if (dims(2) == 1) then
 #if defined(_GPU)
         nsize = product(decomp%ysz)
@@ -176,6 +179,10 @@
      else
         call transpose_y_to_z_complex(src, dst, decomp)
      end if
+
+#ifdef PROFILER
+     if (decomp_profiler_transpose) call decomp_profiler_end("transp_y_z_c")
+#endif
 
   end subroutine transpose_y_to_z_complex_long
 
@@ -193,10 +200,6 @@
 
      integer :: s1, s2, s3, d1, d2, d3
      integer :: ierror
-
-#ifdef PROFILER
-     if (decomp_profiler_transpose) call decomp_profiler_start("transp_y_z_c")
-#endif
 
      s1 = SIZE(src, 1)
      s2 = SIZE(src, 2)
@@ -272,11 +275,6 @@
 
 #endif
 
-#ifdef PROFILER
-     if (decomp_profiler_transpose) call decomp_profiler_end("transp_y_z_c")
-#endif
-
-     return
   end subroutine transpose_y_to_z_complex
 
   subroutine mem_split_yz_real(in, n1, n2, n3, out, iproc, dist, decomp)
