### 2023-09-03 d7ec603
Create visualisation example
