### 2023-09-05 6bc5c5a
visu test output real/imaginary fields for complex tests
