### 2023-09-06 #5790 @p-costa
Count me in. I should mention that I have used a former version of the library for many years, and that I collaborated on a similar library for NVIDIA GPUs, but I feel conformable with reviewing this submission.
