+++ src/transpose_z_to_y.f90
@@ -20,6 +20,7 @@
      real(mytype), dimension(:, :, :), intent(IN) :: src
      real(mytype), dimension(:, :, :), intent(OUT) :: dst
      TYPE(DECOMP_INFO), intent(IN) :: decomp
+
 #if defined(_GPU)
      integer :: istat, nsize
 #endif
@@ -39,7 +40,11 @@
         dst = src
 #endif
      else
-        call transpose_z_to_y_real(src, dst, decomp)
+#if defined(_GPU)
+        call transpose_z_to_y_real(src, dst, decomp, work1_r_d, work2_r_d)
+#else
+        call transpose_z_to_y_real(src, dst, decomp, work1_r, work2_r)
+#endif
      end if
 
 #ifdef PROFILER
@@ -48,13 +53,17 @@
 
   end subroutine transpose_z_to_y_real_long
 
-  subroutine transpose_z_to_y_real(src, dst, decomp)
+  subroutine transpose_z_to_y_real(src, dst, decomp, wk1, wk2)
 
      implicit none
 
      real(mytype), dimension(:, :, :), intent(IN) :: src
      real(mytype), dimension(:, :, :), intent(OUT) :: dst
      TYPE(DECOMP_INFO), intent(IN) :: decomp
+     real(mytype), dimension(:), intent(out) :: wk1, wk2
+#if defined(_GPU)
+     attributes(device) :: wk1, wk2
+#endif
 
 #if defined(_GPU)
      integer :: istat
@@ -73,7 +82,7 @@
      ! rearrange source array as send buffer
 #ifdef EVEN
      if (.not. decomp%even) then
-        call mem_split_zy_real(src, s1, s2, s3, work1_r, dims(2), &
+        call mem_split_zy_real(src, s1, s2, s3, wk1, dims(2), &
                                decomp%z2dist, decomp)
      end if
 #else
@@ -82,7 +91,7 @@
 
 #if defined(_GPU)
      !$acc host_data use_device(src)
-     istat = cudaMemcpy(work1_r_d, src, s1 * s2 * s3, cudaMemcpyDeviceToDevice)
+     istat = cudaMemcpy(wk1, src, s1 * s2 * s3, cudaMemcpyDeviceToDevice)
      !$acc end host_data
      if (istat /= 0) call decomp_2d_abort(__FILE__, __LINE__, istat, "cudaMemcpy2D")
 #endif
@@ -92,48 +101,43 @@
      ! define receive buffer
 #ifdef EVEN
      if (decomp%even) then
-        call MPI_ALLTOALL(src, decomp%z2count, &
-                          real_type, work2_r, decomp%y2count, &
-                          real_type, DECOMP_2D_COMM_ROW, ierror)
+        call MPI_ALLTOALL(src, decomp%z2count, real_type, &
+                          wk2, decomp%y2count, real_type, &
+                          DECOMP_2D_COMM_ROW, ierror)
      else
-        call MPI_ALLTOALL(work1_r, decomp%z2count, &
-                          real_type, work2_r, decomp%y2count, &
-                          real_type, DECOMP_2D_COMM_ROW, ierror)
+        call MPI_ALLTOALL(wk1, decomp%z2count, real_type, &
+                          wk2, decomp%y2count, real_type, &
+                          DECOMP_2D_COMM_ROW, ierror)
      end if
      if (ierror /= 0) call decomp_2d_abort(__FILE__, __LINE__, ierror, "MPI_ALLTOALL")
 #else
 
 #if defined(_GPU)
 #if defined(_NCCL)
-     call decomp_2d_nccl_send_recv_row(work2_r_d, &
-                                       work1_r_d, &
+     call decomp_2d_nccl_send_recv_row(wk2, &
+                                       wk1, &
[…]
