### 2023-09-07 #5813 @editorialbot
[REVIEW]: The new version of the 2DECOMP&FFT library
**Submitting author:** @rfj82982 (<a href="http://orcid.org/0000-0001-6325-7629">Stefano Rolfo</a>)
**Repository:** https://github.com/2decomp-fft/2decomp-fft.git
**Branch with paper.md** (empty if default branch): SubmissionJOSS
**Version:** V2.0.1
