### 2023-09-11 d3b4f9b
Some code is duplicated in the examples. Move it to a dedicated module
