+++ examples/halo_test/halo_test.f90
@@ -12,6 +12,7 @@ program halo_test
    use decomp_2d
    use decomp_2d_constants
    use decomp_2d_mpi
+   use decomp_2d_testing
    use MPI
 #if defined(_GPU)
    use cudafor
@@ -25,8 +26,6 @@ program halo_test
    integer :: p_row = 0, p_col = 0
    integer :: resize_domain
    integer :: nranks_tot
-   integer :: nargin, arg, FNLength, status, DecInd
-   character(len=80) :: InputFN
 
    real(mytype), allocatable, dimension(:, :, :) :: u1, v1, w1
    real(mytype), allocatable, dimension(:, :, :) :: u2, v2, w2
@@ -53,43 +52,11 @@ program halo_test
    ny = ny_base * resize_domain
    nz = nz_base * resize_domain
    ! Now we can check if user put some inputs
-   ! Handle input file like a boss -- GD
-   nargin = command_argument_count()
-   if ((nargin == 0) .or. (nargin == 2) .or. (nargin == 5)) then
-      do arg = 1, nargin
-         call get_command_argument(arg, InputFN, FNLength, status)
-         read (InputFN, *, iostat=status) DecInd
-         if (arg == 1) then
-            p_row = DecInd
-         elseif (arg == 2) then
-            p_col = DecInd
-         elseif (arg == 3) then
-            nx = DecInd
-         elseif (arg == 4) then
-            ny = DecInd
-         elseif (arg == 5) then
-            nz = DecInd
-         end if
-      end do
-   else
-      ! nrank not yet computed we need to avoid write
-      ! for every rank
-      call MPI_COMM_RANK(MPI_COMM_WORLD, nrank, ierror)
-      if (nrank == 0) then
-         print *, "This Test takes no inputs or 2 inputs as"
-         print *, "  1) p_row (default=0)"
-         print *, "  2) p_col (default=0)"
-         print *, "or 5 inputs as"
-         print *, "  1) p_row (default=0)"
-         print *, "  2) p_col (default=0)"
-         print *, "  3) nx "
-         print *, "  4) ny "
-         print *, "  5) nz "
-         print *, "Number of inputs is not correct and the defult settings"
-         print *, "will be used"
-      end if
-   end if
+   call decomp_2d_testing_init(p_row, p_col, nx, ny, nz)
+
    call decomp_2d_init(nx, ny, nz, p_row, p_col)
+
+   call decomp_2d_testing_log()
 
    xlast = xsize(1) - 1
    if (xend(2) == ny) then
