+++ src/CMakeLists.txt
@@ -39,6 +39,10 @@ endif()
 
 set(SRCFILES ${files_decomp} ${files_fft} ${prof_files})
 
+if (${BUILD_TESTING})
+  list(APPEND SRCFILES decomp_2d_testing.f90)
+endif (${BUILD_TESTING})
+
 add_library(decomp2d ${SRCFILES})
 target_include_directories(decomp2d PUBLIC
   $<BUILD_INTERFACE:${CMAKE_Fortran_MODULE_DIRECTORY}>

[…]
