### 2023-09-11 4a71ff5
Skip the submodule to fix compilation with nvhpc
