+++ src/decomp_2d.f90
@@ -303,31 +303,6 @@ module decomp_2d
       end subroutine decomp_profiler_end_char
    end interface decomp_profiler_end
 
-   ! Interpolation submodule
-   interface d2d_interp_var3d
-      module subroutine d2d_interp_var3d_0(ipencil, varin, decompin, varout, decompout, interp)
-         integer, intent(in) :: ipencil
-         type(decomp_info), intent(in) :: decompin, decompout
-         real(mytype), dimension(:, :, :), intent(in) :: varin
-         real(mytype), dimension(:, :, :), intent(out) :: varout
-         integer, intent(in), optional :: interp
-      end subroutine d2d_interp_var3d_0
-      module subroutine d2d_interp_var3d_1(ipencil, varin, varout, decompout, interp)
-         integer, intent(in) :: ipencil
-         type(decomp_info), intent(in) :: decompout
-         real(mytype), dimension(:, :, :), intent(in) :: varin
-         real(mytype), dimension(:, :, :), intent(out) :: varout
-         integer, intent(in), optional :: interp
-      end subroutine d2d_interp_var3d_1
-      module subroutine d2d_interp_var3d_2(ipencil, varin, decompin, varout, interp)
-         integer, intent(in) :: ipencil
-         type(decomp_info), intent(in) :: decompin
-         real(mytype), dimension(:, :, :), intent(in) :: varin
-         real(mytype), dimension(:, :, :), intent(out) :: varout
-         integer, intent(in), optional :: interp
-      end subroutine d2d_interp_var3d_2
-   end interface d2d_interp_var3d
-
 contains
 
 #include "decomp_2d_init_fin.f90"
