+++ src/io_family.f90
@@ -1,0 +1,532 @@
+!=======================================================================
+! This is part of the 2DECOMP&FFT library
+!
+! 2DECOMP&FFT is a software framework for general-purpose 2D (pencil)
+! decomposition. It also implements a highly scalable distributed
+! three-dimensional Fast Fourier Transform (FFT).
+!
+! Copyright (C) 2009-2012 Ning Li, the Numerical Algorithms Group (NAG)
+! Copyright (C) 2021               the University of Edinburgh (UoE)
+!
+!=======================================================================
+
+!
+! This is the IO module for families of readers / writers
+!    => Initialize / finalize a family of readers / writers
+!    => Register a variable in a family of readers / writers
+!
+
+module decomp_2d_io_family
+
+   use decomp_2d_constants
+   use decomp_2d_mpi
+   use decomp_2d
+#ifdef ADIOS2
+   use adios2
+#endif
+   use, intrinsic :: iso_fortran_env, only: real32, real64
+
+   implicit none
+
+#ifdef ADIOS2
+   type(adios2_adios), pointer :: default_adios => null()
+#endif
+
+   !
+   ! derived type to store info for a family of readers / writers
+   !
+   type, public :: d2d_io_family
+      integer :: type = decomp_2d_io_none                     ! Type of the writer
+      character(:), allocatable :: label                      ! Label of the writer
+#ifdef ADIOS2
+      type(adios2_adios), pointer :: adios                    ! adios2 only
+      type(adios2_io) :: io                                   ! adios2 only
+#endif
+   contains
+      procedure :: init => d2d_io_family_init                 ! ADIOS2 writer if possible, MPI otherwise
+      procedure :: mpi_init => d2d_io_family_mpi_init         ! Force MPI writer
+      procedure :: adios2_init => d2d_io_family_adios2_init   ! Force ADIOS2 writer
+      procedure :: fin => d2d_io_family_fin                   ! Clear the writer
+      procedure :: register_var => d2d_io_family_register_var ! Register a variable
+   end type d2d_io_family
+
+   private
+
+#ifdef ADIOS2
+   public :: decomp_2d_io_family_get_default_adios, &
+             decomp_2d_io_family_set_default_adios
+#endif
+
+contains
+
+   !
+   ! Initialize a new family of readers / writers (default type)
+   !
+   subroutine d2d_io_family_init(family, label)
+
+      implicit none
+
+      class(d2d_io_family), intent(inout) :: family
+      character(len=*), intent(in) :: label
+
+#ifdef ADIOS2
+      call family%adios2_init(label)
+#else
+      call family%mpi_init(label)
+#endif
+
+   end subroutine d2d_io_family_init
+
+   !
+   ! Initialize a new family of MPI readers / writers
+   !
+   subroutine d2d_io_family_mpi_init(family, label)
+
+      implicit none
+
+      class(d2d_io_family), intent(inout) :: family
+      character(len=*), intent(in) :: label
+
+#ifdef PROFILER
+      if (decomp_profiler_io) call decomp_profiler_start("d2d_io_family_mpi_init")
+#endif
+
+      ! Safety check
+      if (family%type /= decomp_2d_io_none) then
+         call decomp_2d_abort(__FILE__, __LINE__, -1, "Family was not cleared "//label)
+      end if
+
+      family%type = decomp_2d_io_mpi
+      family%label = label
+
+#ifdef PROFILER
+      if (decomp_profiler_io) call decomp_profiler_end("d2d_io_family_mpi_init")
+#endif
+
+   end subroutine d2d_io_family_mpi_init
+
+   !
+   ! Initialize a new family of ADIOS2 readers / writers
+   !
[…]
