+++ examples/fft_physical_z/CMakeLists.txt
@@ -1,5 +1,5 @@
-file(GLOB files_fft_c2c fft_c2c_z.f90)
-file(GLOB files_fft_r2c fft_r2c_z.f90)
+file(GLOB files_fft_c2c fft_c2c_z.f90 ../utilities.f90)
+file(GLOB files_fft_r2c fft_r2c_z.f90 ../utilities.f90)
 
 include_directories(${CMAKE_SOURCE_DIR}/src)
 
