### 2023-09-11 9cfcad4
Avoid submodules for the reader / writer IO object
