+++ src/io_object.f90
@@ -91,9 +91,12 @@
 !       should allow reduced precision operations easily
 !
 
-submodule(decomp_2d_io) decomp_2d_io_objects
+module decomp_2d_io_object
 
    use decomp_2d_constants
+   use decomp_2d_io_family
+   use decomp_2d_mpi
+   use decomp_2d
 #ifdef ADIOS2
    use adios2
 #endif
@@ -101,272 +104,29 @@ submodule(decomp_2d_io) decomp_2d_io_objects
 
    implicit none
 
+   ! derived type to store info for a reader / writer
+   type, public :: d2d_io
+      type(d2d_io_family), pointer :: family        ! Associated family
+      character(:), allocatable :: label            ! Label of the writer
+      logical :: is_open = .false.                  ! True if the writer is open
+#ifdef ADIOS2
+      logical :: is_active = .false.                ! True if the writer is active (adios2 only)
+      type(adios2_engine) :: engine                 ! adios2 only (only one engine / writer currently)
+#endif
+      integer :: fh                                 ! File handle (mpi only)
+      integer(kind=MPI_OFFSET_KIND) :: disp         ! Displacement offset (mpi only)
+   contains
+      procedure :: open => d2d_io_open              ! Open the IO
+      procedure :: start => d2d_io_start            ! Start the IO
+      procedure :: open_start => d2d_io_open_start  ! Open and start the IO
+      procedure :: end => d2d_io_end                ! End the IO
+      procedure :: close => d2d_io_close            ! Close the IO
+      procedure :: end_close => d2d_io_end_close    ! End and close the IO
+   end type d2d_io
+
+   private
+
 contains
-
-   !
-   ! Initialize a new family of readers / writers (default type)
-   !
-   subroutine d2d_io_family_init(family, label)
-
-      implicit none
-
-      class(d2d_io_family), intent(inout) :: family
-      character(len=*), intent(in) :: label
-
-#ifdef ADIOS2
-      call family%adios2_init(label)
-#else
-      call family%mpi_init(label)
-#endif
-
-   end subroutine d2d_io_family_init
-
-   !
-   ! Initialize a new family of MPI readers / writers
-   !
-   subroutine d2d_io_family_mpi_init(family, label)
-
-      implicit none
-
-      class(d2d_io_family), intent(inout) :: family
-      character(len=*), intent(in) :: label
-
-#ifdef PROFILER
-      if (decomp_profiler_io) call decomp_profiler_start("d2d_io_family_mpi_init")
-#endif
-
-      ! Safety check
-      if (family%type /= decomp_2d_io_none) then
-         call decomp_2d_abort(__FILE__, __LINE__, -1, "Family was not cleared "//label)
-      end if
-
-      family%type = decomp_2d_io_mpi
-      family%label = label
-
-#ifdef PROFILER
-      if (decomp_profiler_io) call decomp_profiler_end("d2d_io_family_mpi_init")
-#endif
-
-   end subroutine d2d_io_family_mpi_init
-
-   !
-   ! Initialize a new family of ADIOS2 readers / writers
-   !
-   subroutine d2d_io_family_adios2_init(family, label)
-
-      implicit none
-
-      class(d2d_io_family), intent(inout) :: family
-      character(len=*), intent(in) :: label
-
-#ifdef ADIOS2
-      integer :: ierror
-#endif
-
-#ifdef PROFILER
-      if (decomp_profiler_io) call decomp_profiler_start("d2d_io_family_adios2_init")
-#endif
-
-      ! Safety check
-      if (family%type /= decomp_2d_io_none) then
[…]
