+++ src/CMakeLists.txt
@@ -4,8 +4,8 @@ file(GLOB files_decomp factor.f90
 	               decomp_2d.f90
 		       log.f90
 		       io.f90
-		       io_objects.f90
 		       io_family.f90
+  		       io_object.f90
 		       interp.f90)
 include_directories(${CMAKE_SOURCE_DIR}/src)
 
