+++ src/io_family.f90
@@ -21,6 +21,7 @@ module decomp_2d_io_family
    use decomp_2d_constants
    use decomp_2d_mpi
    use decomp_2d
+   use MPI
 #ifdef ADIOS2
    use adios2
 #endif
@@ -43,11 +44,12 @@ module decomp_2d_io_family
       type(adios2_io) :: io                                   ! adios2 only
 #endif
    contains
-      procedure :: init => d2d_io_family_init                 ! ADIOS2 writer if possible, MPI otherwise
-      procedure :: mpi_init => d2d_io_family_mpi_init         ! Force MPI writer
-      procedure :: adios2_init => d2d_io_family_adios2_init   ! Force ADIOS2 writer
-      procedure :: fin => d2d_io_family_fin                   ! Clear the writer
-      procedure :: register_var => d2d_io_family_register_var ! Register a variable
+      procedure :: init => d2d_io_family_init                     ! ADIOS2 writer if possible, MPI otherwise
+      procedure :: mpi_init => d2d_io_family_mpi_init             ! Force MPI writer
+      procedure :: adios2_init => d2d_io_family_adios2_init       ! Force ADIOS2 writer
+      procedure :: fin => d2d_io_family_fin                       ! Clear the writer
+      procedure :: register_var3d => d2d_io_family_register_var3d ! Register a 3D variable
+      procedure :: register_var2d => d2d_io_family_register_var2d ! Register a 2D variable
    end type d2d_io_family
 
    private
@@ -210,125 +212,263 @@ contains
    end subroutine d2d_io_family_fin
 
    !
-   ! Register one variable for the given family of readers / writers
-   !    1 <= ipencil <= 3
-   !    0 <= iplane <= 3
+   ! High-level. Register one 3D variable for the given family of readers / writers.
    !
-   module subroutine d2d_io_family_register_var(family, &
-                                                varname, &
-                                                ipencil, &
-                                                iplane, &
-                                                type, &
-                                                opt_decomp, &
-                                                opt_nplanes)
+   !    1 <= ipencil <= 3
+   !    type can be MPI_REAL
+   !                MPI_DOUBLE_PRECISION
+   !                MPI_COMPLEX
+   !                MPI_DOUBLE_COMPLEX
+   !
+   subroutine d2d_io_family_register_var3d(family, &
+                                           varname, &
+                                           ipencil, &
+                                           type, &
+                                           opt_decomp)
 
       implicit none
 
       class(d2d_io_family), intent(inout) :: family
       character(len=*), intent(in) :: varname
       integer, intent(in) :: ipencil ! (x-pencil=1; y-pencil=2; z-pencil=3)
-      integer, intent(in) :: iplane
       integer, intent(in) :: type
       type(decomp_info), intent(in), optional :: opt_decomp
-      integer, intent(in), optional :: opt_nplanes
-
-#ifdef ADIOS2
-      integer :: nplanes
-      type(adios2_variable) :: var_handle
-      integer, dimension(3) :: sizes, subsizes, starts
-      integer, parameter :: ndims = 3
-      logical, parameter :: adios2_constant_dims = .true.
-      integer :: data_type
-      integer :: ierror
-#endif
 
 #ifdef PROFILER
-      if (decomp_profiler_io) call decomp_profiler_start("d2d_io_family_register_var")
+      if (decomp_profiler_io) call decomp_profiler_start("d2d_io_family_register_var3d")
 #endif
+
+      if (present(opt_decomp)) then
[…]
