+++ src/interp.f90
@@ -26,27 +26,27 @@ module decomp_2d_interp
 
    ! Interpolation subroutines
    interface decomp_2d_interp_var3d
-      module subroutine d2d_interp_var3d_0(ipencil, varin, decompin, varout, decompout, interp)      
+      module subroutine d2d_interp_var3d_0(ipencil, varin, decompin, varout, decompout, interp)
          integer, intent(in) :: ipencil
          type(decomp_info), intent(in) :: decompin, decompout
          real(mytype), dimension(:, :, :), intent(in) :: varin
-         real(mytype), dimension(:, :, :), intent(out) :: varout                                     
-         integer, intent(in), optional :: interp                                                     
+         real(mytype), dimension(:, :, :), intent(out) :: varout
+         integer, intent(in), optional :: interp
       end subroutine d2d_interp_var3d_0
-      module subroutine d2d_interp_var3d_1(ipencil, varin, varout, decompout, interp)                
+      module subroutine d2d_interp_var3d_1(ipencil, varin, varout, decompout, interp)
          integer, intent(in) :: ipencil
          type(decomp_info), intent(in) :: decompout
          real(mytype), dimension(:, :, :), intent(in) :: varin
-         real(mytype), dimension(:, :, :), intent(out) :: varout                                     
-         integer, intent(in), optional :: interp                                                     
+         real(mytype), dimension(:, :, :), intent(out) :: varout
+         integer, intent(in), optional :: interp
       end subroutine d2d_interp_var3d_1
-      module subroutine d2d_interp_var3d_2(ipencil, varin, decompin, varout, interp)                 
+      module subroutine d2d_interp_var3d_2(ipencil, varin, decompin, varout, interp)
          integer, intent(in) :: ipencil
          type(decomp_info), intent(in) :: decompin
          real(mytype), dimension(:, :, :), intent(in) :: varin
-         real(mytype), dimension(:, :, :), intent(out) :: varout                                     
-         integer, intent(in), optional :: interp                                                     
-      end subroutine d2d_interp_var3d_2                                                              
+         real(mytype), dimension(:, :, :), intent(out) :: varout
+         integer, intent(in), optional :: interp
+      end subroutine d2d_interp_var3d_2
    end interface decomp_2d_interp_var3d
 
 contains
