### 2023-09-12 554c2bd
Link the examples utilities as an object library (#265)
CMake object libraries are "logical" libraries that allow us to
define a set of object files that depend on another library, here

example_utils depends on decomp2d
example/X depends on decomp2d and example_utils
