+++ examples/fft_physical_z/CMakeLists.txt
@@ -1,13 +1,13 @@
-file(GLOB files_fft_c2c fft_c2c_z.f90 ../utilities.f90)
-file(GLOB files_fft_r2c fft_r2c_z.f90 ../utilities.f90)
+file(GLOB files_fft_c2c fft_c2c_z.f90)
+file(GLOB files_fft_r2c fft_r2c_z.f90)
 
 include_directories(${CMAKE_SOURCE_DIR}/src)
 
 add_executable(fft_c2c_z ${files_fft_c2c})
 add_executable(fft_r2c_z ${files_fft_r2c})
 
-target_link_libraries(fft_c2c_z PRIVATE decomp2d)
-target_link_libraries(fft_r2c_z PRIVATE decomp2d)
+target_link_libraries(fft_c2c_z PRIVATE decomp2d examples_utils)
+target_link_libraries(fft_r2c_z PRIVATE decomp2d examples_utils)
 
 # Run the test(s)
 set(run_dir "${test_dir}/fft_c2c_z")
