### 2023-09-12 6a55673
Add a log subroutine in the object d2d_io_family
