### 2023-09-12 8811e4c
Fix halo test, use local not global variable (#267)
