### 2023-09-12 e45e7de
Some code is duplicated in the examples. Move it to a dedicated module (#264)
* Some code is duplicated in the examples. Move it to a dedicated module

* Move decomp_2d_testing to the example folder

* Minor fix

* Link the examples utilities as an object library (#265)

CMake object libraries are "logical" libraries that allow us to
define a set of object files that depend on another library, here

example_utils depends on decomp2d
example/X depends on decomp2d and example_utils

---------

Co-authored-by: Paul Bartholomew <p.bartholomew@epcc.ed.ac.uk>
