### 2023-09-12 be7f23c
IO : remove decomp_2d_init_io and rename decomp_2d_io_finalise
