### 2023-09-12 92cbde1
IO : add a reduced precision when registering a variable
