+++ src/io.f90
@@ -199,23 +199,35 @@ contains
    !
    ! High-level. Using MPI-IO / ADIOS2 to write a 3D array to a file
    !
-   subroutine write_one_freal(ipencil, var, io, varname, &
-                              opt_mode, opt_reduce_prec, opt_decomp)
+   subroutine write_one_freal(ipencil, var, varname, opt_mode, opt_family, &
+                              opt_io, opt_dirname, opt_reduce_prec, opt_decomp)
 
       implicit none
 
       integer, intent(IN) :: ipencil
       real(kind(0._real32)), contiguous, dimension(:, :, :), intent(IN) :: var
-      type(d2d_io), intent(inout) :: io
       character(len=*), intent(in) :: varname
       integer, intent(in), optional :: opt_mode
+      type(d2d_io_family), intent(in), optional :: opt_family
+      type(d2d_io), intent(inout), optional :: opt_io
+      character(len=*), intent(in), optional :: opt_dirname
       logical, intent(in), optional :: opt_reduce_prec
       TYPE(DECOMP_INFO), intent(IN), optional :: opt_decomp
 
       if (present(opt_decomp)) then
-         call write_one(ipencil, io, varname, opt_decomp, opt_mode, freal=var)
+         call write_one(ipencil, varname, opt_decomp, &
+                        opt_mode=opt_mode, &
+                        opt_family=opt_family, &
+                        opt_io=opt_io, &
+                        opt_dirname=opt_dirname, &
+                        freal=var)
       else
-         call write_one(ipencil, io, varname, decomp_main, opt_mode, freal=var)
+         call write_one(ipencil, varname, decomp_main, &
+                        opt_mode=opt_mode, &
+                        opt_family=opt_family, &
+                        opt_io=opt_io, &
+                        opt_dirname=opt_dirname, &
+                        freal=var)
       end if
 
       associate (p => opt_reduce_prec)
@@ -223,23 +235,35 @@ contains
 
    end subroutine write_one_freal
    !
-   subroutine write_one_fcplx(ipencil, var, io, varname, &
-                              opt_mode, opt_reduce_prec, opt_decomp)
+   subroutine write_one_fcplx(ipencil, var, varname, opt_mode, opt_family, &
+                              opt_io, opt_dirname, opt_reduce_prec, opt_decomp)
 
       implicit none
 
       integer, intent(IN) :: ipencil
       complex(kind(0._real32)), contiguous, dimension(:, :, :), intent(IN) :: var
-      type(d2d_io), intent(inout) :: io
       character(len=*), intent(in) :: varname
       integer, intent(in), optional :: opt_mode
+      type(d2d_io_family), intent(in), optional :: opt_family
+      type(d2d_io), intent(inout), optional :: opt_io
+      character(len=*), intent(in), optional :: opt_dirname
       logical, intent(in), optional :: opt_reduce_prec
       TYPE(DECOMP_INFO), intent(IN), optional :: opt_decomp
 
       if (present(opt_decomp)) then
-         call write_one(ipencil, io, varname, opt_decomp, opt_mode, fcplx=var)
+         call write_one(ipencil, varname, opt_decomp, &
+                        opt_mode=opt_mode, &
+                        opt_family=opt_family, &
+                        opt_io=opt_io, &
+                        opt_dirname=opt_dirname, &
+                        fcplx=var)
       else
-         call write_one(ipencil, io, varname, decomp_main, opt_mode, fcplx=var)
+         call write_one(ipencil, varname, decomp_main, &
+                        opt_mode=opt_mode, &
+                        opt_family=opt_family, &
+                        opt_io=opt_io, &
+                        opt_dirname=opt_dirname, &
[…]
