+++ examples/halo_test/halo_test.f90
@@ -36,7 +36,6 @@ program halo_test
 
    integer, allocatable, dimension(:) :: seed
 
-   real(mytype) :: err
    integer :: xlast, ylast, zlast
 
    integer :: nx_expected, ny_expected, nz_expected
@@ -509,15 +508,15 @@ contains
       allocate (tmp(size(divh, 1), size(divh, 2), size(divh, 3)))
 
       !$acc kernels default(present)
-      tmp(2:xlast, 2:ylast, 2:zlast) = divh(2:xlast, 2:ylast, 2:zlast) - div1(2:xlast, 2:ylast, 2:zlast)
+      tmp(2:xlast, 2:ylast, 2:zlast) = divh(2:xlast, 2:ylast, 2:zlast) - div(2:xlast, 2:ylast, 2:zlast)
       !$acc end kernels
       error = mag(tmp)
       !$acc kernels default(present)
-      tmp(2:xlast, 2:ylast, 2:zlast) = div1(2:xlast, 2:ylast, 2:zlast)
+      tmp(2:xlast, 2:ylast, 2:zlast) = div(2:xlast, 2:ylast, 2:zlast)
       !$acc end kernels
       divmag = mag(tmp)
 
-      if (err < epsilon(divmag) * divmag) then
+      if (error < epsilon(divmag) * divmag) then
          passing = .true.
       else
          passing = .false.
@@ -530,7 +529,8 @@ contains
 #ifdef DEBUG
          write (*, *) (divh(i, i, i), i=2, 13)
 #endif
-         write (*, *) 'Error: ', err, '; Relative: ', err / divmag
+         print *, error, divmag
+         write (*, *) 'Error: ', error, '; Relative: ', error / divmag
          write (*, *) 'Pass: ', passing
       end if
       deallocate (tmp)
