+++ src/io.f90
@@ -60,8 +60,10 @@ module decomp_2d_io
    end interface decomp_2d_write_one
 
    interface decomp_2d_read_one
-      module procedure read_one_real
-      module procedure read_one_complex
+      module procedure read_one_freal
+      module procedure read_one_dreal
+      module procedure read_one_fcplx
+      module procedure read_one_dcplx
    end interface decomp_2d_read_one
 
    interface decomp_2d_write_var
@@ -390,6 +392,125 @@ contains
    end subroutine write_one_dcplx
 
    !
+   ! High-level. Using MPI-IO / ADIOS2 to read a 3D array from a file
+   !
+   subroutine read_one_freal(ipencil, var, varname, opt_family, &
+                             opt_io, opt_dirname, opt_decomp)
+
+      implicit none
+
+      integer, intent(IN) :: ipencil
+      real(kind(0._real32)), contiguous, dimension(:, :, :), intent(out) :: var
+      character(len=*), intent(in) :: varname
+      type(d2d_io_family), intent(in), optional :: opt_family
+      type(d2d_io), intent(inout), optional :: opt_io
+      character(len=*), intent(in), optional :: opt_dirname
+      TYPE(DECOMP_INFO), intent(IN), optional :: opt_decomp
+
+      if (present(opt_decomp)) then
+         call read_one(ipencil, varname, opt_decomp, &
+                       opt_family=opt_family, &
+                       opt_io=opt_io, &
+                       opt_dirname=opt_dirname, &
+                       freal=var)
+      else
+         call read_one(ipencil, varname, decomp_main, &
+                       opt_family=opt_family, &
+                       opt_io=opt_io, &
+                       opt_dirname=opt_dirname, &
+                       freal=var)
+      end if
+
+   end subroutine read_one_freal
+   !
+   subroutine read_one_fcplx(ipencil, var, varname, opt_family, &
+                             opt_io, opt_dirname, opt_decomp)
+
+      implicit none
+
+      integer, intent(IN) :: ipencil
+      complex(kind(0._real32)), contiguous, dimension(:, :, :), intent(out) :: var
+      character(len=*), intent(in) :: varname
+      type(d2d_io_family), intent(in), optional :: opt_family
+      type(d2d_io), intent(inout), optional :: opt_io
+      character(len=*), intent(in), optional :: opt_dirname
+      TYPE(DECOMP_INFO), intent(IN), optional :: opt_decomp
+
+      if (present(opt_decomp)) then
+         call read_one(ipencil, varname, opt_decomp, &
+                       opt_family=opt_family, &
+                       opt_io=opt_io, &
+                       opt_dirname=opt_dirname, &
+                       fcplx=var)
+      else
+         call read_one(ipencil, varname, decomp_main, &
+                       opt_family=opt_family, &
+                       opt_io=opt_io, &
+                       opt_dirname=opt_dirname, &
+                       fcplx=var)
+      end if
+
+   end subroutine read_one_fcplx
+   !
+   subroutine read_one_dreal(ipencil, var, varname, opt_family, &
+                             opt_io, opt_dirname, opt_decomp)
+
+      implicit none
+
+      integer, intent(IN) :: ipencil
+      real(kind(0._real64)), contiguous, dimension(:, :, :), intent(out) :: var
+      character(len=*), intent(in) :: varname
+      type(d2d_io_family), intent(in), optional :: opt_family
[…]
