### 2023-09-12 b29c30a
Add a subroutine for open and start code in read_one and write_one
