### 2023-09-12 a6ca8f1
Add a dedicated subroutine to update the MPI displacement
