### 2023-09-13 3785930
Fuse mpi_write and mpi_read
