### 2023-09-13 751cd41
Extend x allocator to deal with single and double precision
