### 2023-09-13 e9426c0
Extend y allocator to deal with single and double precision
