+++ src/alloc.f90
@@ -149,7 +149,6 @@
                              'Memory allocation failed when creating new arrays')
      end if
 
-     return
   end subroutine alloc_x_fcplx
 
   subroutine alloc_x_dcplx(var, decomp, opt_global)
@@ -184,26 +183,36 @@
                              'Memory allocation failed when creating new arrays')
      end if
 
-     return
   end subroutine alloc_x_dcplx
 
   ! Y-pencil real arrays
-  subroutine alloc_y_real_short(var, opt_global)
+  subroutine alloc_y_freal_short(var, opt_global)
 
      implicit none
 
-     real(mytype), allocatable, dimension(:, :, :) :: var
+     real(kind(0._real32)), allocatable, dimension(:, :, :) :: var
      logical, intent(IN), optional :: opt_global
 
      call alloc_y(var, decomp_main, opt_global)
 
-  end subroutine alloc_y_real_short
+  end subroutine alloc_y_freal_short
 
-  subroutine alloc_y_real(var, decomp, opt_global)
+  subroutine alloc_y_dreal_short(var, opt_global)
 
      implicit none
 
-     real(mytype), allocatable, dimension(:, :, :) :: var
+     real(kind(0._real64)), allocatable, dimension(:, :, :) :: var
+     logical, intent(IN), optional :: opt_global
+
+     call alloc_y(var, decomp_main, opt_global)
+
+  end subroutine alloc_y_dreal_short
+
+  subroutine alloc_y_freal(var, decomp, opt_global)
+
+     implicit none
+
+     real(kind(0._real32)), allocatable, dimension(:, :, :) :: var
      TYPE(DECOMP_INFO), intent(IN) :: decomp
      logical, intent(IN), optional :: opt_global
 
@@ -213,7 +222,7 @@
      if (present(opt_global)) then
         global = opt_global
      else
-        global = .false.
+        global = default_opt_global
      end if
 
      if (global) then
@@ -231,26 +240,70 @@
                              'Memory allocation failed when creating new arrays')
      end if
 
-     return
-  end subroutine alloc_y_real
+  end subroutine alloc_y_freal
+
+  subroutine alloc_y_dreal(var, decomp, opt_global)
+
+     implicit none
+
+     real(kind(0._real64)), allocatable, dimension(:, :, :) :: var
+     TYPE(DECOMP_INFO), intent(IN) :: decomp
+     logical, intent(IN), optional :: opt_global
+
+     logical :: global
+     integer :: alloc_stat, errorcode
+
+     if (present(opt_global)) then
+        global = opt_global
+     else
+        global = default_opt_global
+     end if
+
+     if (global) then
+        allocate (var(decomp%yst(1):decomp%yen(1), &
+                      decomp%yst(2):decomp%yen(2), decomp%yst(3):decomp%yen(3)), &
+                  stat=alloc_stat)
+     else
+        allocate (var(decomp%ysz(1), decomp%ysz(2), decomp%ysz(3)), &
+                  stat=alloc_stat)
+     end if
+
+     if (alloc_stat /= 0) then
+        errorcode = 8
+        call decomp_2d_abort(errorcode, &
+                             'Memory allocation failed when creating new arrays')
+     end if
+
+  end subroutine alloc_y_dreal
 
   ! Y-pencil complex arrays
-  subroutine alloc_y_complex_short(var, opt_global)
+  subroutine alloc_y_fcplx_short(var, opt_global)
 
      implicit none
 
-     complex(mytype), allocatable, dimension(:, :, :) :: var
+     complex(kind(0._real32)), allocatable, dimension(:, :, :) :: var
      logical, intent(IN), optional :: opt_global
 
[…]
