### 2023-09-13 7d2a008
Extend z allocator to deal with single and double precision
