### 2023-09-13 735da8e
Assume target attribute on argument opt_decomp to avoid duplicated code
