+++ src/io.f90
@@ -19,7 +19,7 @@ module decomp_2d_io
 
    implicit none
 
-   ! Default reduced precision for write_one
+   ! Default reduced precision for read_one and write_one
    logical, parameter :: default_opt_reduce_prec = .false.
 
    ! Default IO family of readers / writers
@@ -217,23 +217,24 @@ contains
       type(d2d_io), intent(inout), optional :: opt_io
       character(len=*), intent(in), optional :: opt_dirname
       logical, intent(in), optional :: opt_reduce_prec
-      TYPE(DECOMP_INFO), intent(IN), optional :: opt_decomp
+      TYPE(DECOMP_INFO), target, intent(IN), optional :: opt_decomp
+
+      TYPE(DECOMP_INFO), pointer :: decomp
 
       if (present(opt_decomp)) then
-         call write_one(ipencil, varname, opt_decomp, &
-                        opt_mode=opt_mode, &
-                        opt_family=opt_family, &
-                        opt_io=opt_io, &
-                        opt_dirname=opt_dirname, &
-                        freal=var)
+         decomp => opt_decomp
       else
-         call write_one(ipencil, varname, decomp_main, &
-                        opt_mode=opt_mode, &
-                        opt_family=opt_family, &
-                        opt_io=opt_io, &
-                        opt_dirname=opt_dirname, &
-                        freal=var)
+         decomp => decomp_main
       end if
+
+      call write_one(ipencil, varname, decomp, &
+                     opt_mode=opt_mode, &
+                     opt_family=opt_family, &
+                     opt_io=opt_io, &
+                     opt_dirname=opt_dirname, &
+                     freal=var)
+
+      nullify (decomp)
 
       associate (p => opt_reduce_prec)
       end associate
@@ -253,23 +254,24 @@ contains
       type(d2d_io), intent(inout), optional :: opt_io
       character(len=*), intent(in), optional :: opt_dirname
       logical, intent(in), optional :: opt_reduce_prec
-      TYPE(DECOMP_INFO), intent(IN), optional :: opt_decomp
+      TYPE(DECOMP_INFO), target, intent(IN), optional :: opt_decomp
+
+      TYPE(DECOMP_INFO), pointer :: decomp
 
       if (present(opt_decomp)) then
-         call write_one(ipencil, varname, opt_decomp, &
-                        opt_mode=opt_mode, &
-                        opt_family=opt_family, &
-                        opt_io=opt_io, &
-                        opt_dirname=opt_dirname, &
-                        fcplx=var)
+         decomp => opt_decomp
       else
-         call write_one(ipencil, varname, decomp_main, &
-                        opt_mode=opt_mode, &
-                        opt_family=opt_family, &
-                        opt_io=opt_io, &
-                        opt_dirname=opt_dirname, &
-                        fcplx=var)
+         decomp => decomp_main
       end if
+
+      call write_one(ipencil, varname, decomp, &
+                     opt_mode=opt_mode, &
+                     opt_family=opt_family, &
+                     opt_io=opt_io, &
+                     opt_dirname=opt_dirname, &
+                     fcplx=var)
+
+      nullify (decomp)
 
       associate (p => opt_reduce_prec)
       end associate
@@ -289,9 +291,16 @@ contains
       type(d2d_io), intent(inout), optional :: opt_io
       character(len=*), intent(in), optional :: opt_dirname
       logical, intent(in), optional :: opt_reduce_prec
-      TYPE(DECOMP_INFO), intent(IN), optional :: opt_decomp
+      TYPE(DECOMP_INFO), target, intent(IN), optional :: opt_decomp
 
       logical :: reduce
+      TYPE(DECOMP_INFO), pointer :: decomp
+
+      if (present(opt_decomp)) then
+         decomp => opt_decomp
+      else
[…]
