### 2023-09-15 16a19bb
Open and start IO object without providing a family
