### 2023-09-15 f789361
Example of IO without providing a dedicated IO family
