### 2023-09-22 1168095
Simpler declaration of the variable precision
