+++ src/alloc.f90
@@ -8,7 +8,7 @@
 
      implicit none
 
-     real(kind(0._real32)), allocatable, dimension(:, :, :) :: var
+     real(real32), allocatable, dimension(:, :, :) :: var
      logical, intent(IN), optional :: opt_global
 
      call alloc_x(var, decomp_main, opt_global)
@@ -19,7 +19,7 @@
 
      implicit none
 
-     real(kind(0._real64)), allocatable, dimension(:, :, :) :: var
+     real(real64), allocatable, dimension(:, :, :) :: var
      logical, intent(IN), optional :: opt_global
 
      call alloc_x(var, decomp_main, opt_global)
@@ -30,7 +30,7 @@
 
      implicit none
 
-     real(kind(0._real32)), allocatable, dimension(:, :, :) :: var
+     real(real32), allocatable, dimension(:, :, :) :: var
      TYPE(DECOMP_INFO), intent(IN) :: decomp
      logical, intent(IN), optional :: opt_global
 
@@ -64,7 +64,7 @@
 
      implicit none
 
-     real(kind(0._real64)), allocatable, dimension(:, :, :) :: var
+     real(real64), allocatable, dimension(:, :, :) :: var
      TYPE(DECOMP_INFO), intent(IN) :: decomp
      logical, intent(IN), optional :: opt_global
 
@@ -99,7 +99,7 @@
 
      implicit none
 
-     complex(kind(0._real32)), allocatable, dimension(:, :, :) :: var
+     complex(real32), allocatable, dimension(:, :, :) :: var
      logical, intent(IN), optional :: opt_global
 
      call alloc_x(var, decomp_main, opt_global)
@@ -110,7 +110,7 @@
 
      implicit none
 
-     complex(kind(0._real64)), allocatable, dimension(:, :, :) :: var
+     complex(real64), allocatable, dimension(:, :, :) :: var
      logical, intent(IN), optional :: opt_global
 
      call alloc_x(var, decomp_main, opt_global)
@@ -121,7 +121,7 @@
 
      implicit none
 
-     complex(kind(0._real32)), allocatable, dimension(:, :, :) :: var
+     complex(real32), allocatable, dimension(:, :, :) :: var
      TYPE(DECOMP_INFO), intent(IN) :: decomp
      logical, intent(IN), optional :: opt_global
 
@@ -155,7 +155,7 @@
 
      implicit none
 
-     complex(kind(0._real64)), allocatable, dimension(:, :, :) :: var
+     complex(real64), allocatable, dimension(:, :, :) :: var
      TYPE(DECOMP_INFO), intent(IN) :: decomp
      logical, intent(IN), optional :: opt_global
 
@@ -190,7 +190,7 @@
 
      implicit none
 
-     real(kind(0._real32)), allocatable, dimension(:, :, :) :: var
+     real(real32), allocatable, dimension(:, :, :) :: var
      logical, intent(IN), optional :: opt_global
 
      call alloc_y(var, decomp_main, opt_global)
@@ -201,7 +201,7 @@
 
      implicit none
 
-     real(kind(0._real64)), allocatable, dimension(:, :, :) :: var
+     real(real64), allocatable, dimension(:, :, :) :: var
      logical, intent(IN), optional :: opt_global
 
      call alloc_y(var, decomp_main, opt_global)
@@ -212,7 +212,7 @@
 
      implicit none
 
-     real(kind(0._real32)), allocatable, dimension(:, :, :) :: var
+     real(real32), allocatable, dimension(:, :, :) :: var
      TYPE(DECOMP_INFO), intent(IN) :: decomp
      logical, intent(IN), optional :: opt_global
 
@@ -246,7 +246,7 @@
 
      implicit none
[…]
