+++ src/io.f90
@@ -250,7 +250,7 @@ contains
       implicit none
 
       integer, intent(IN) :: ipencil
-      real(kind(0._real32)), contiguous, dimension(:, :, :), intent(IN) :: var
+      real(real32), contiguous, dimension(:, :, :), intent(IN) :: var
       character(len=*), intent(in) :: varname
       integer, intent(in), optional :: opt_mode
       type(d2d_io_family), target, intent(in), optional :: opt_family
@@ -288,7 +288,7 @@ contains
       implicit none
 
       integer, intent(IN) :: ipencil
-      complex(kind(0._real32)), contiguous, dimension(:, :, :), intent(IN) :: var
+      complex(real32), contiguous, dimension(:, :, :), intent(IN) :: var
       character(len=*), intent(in) :: varname
       integer, intent(in), optional :: opt_mode
       type(d2d_io_family), target, intent(in), optional :: opt_family
@@ -326,7 +326,7 @@ contains
       implicit none
 
       integer, intent(IN) :: ipencil
-      real(kind(0._real64)), contiguous, dimension(:, :, :), intent(IN) :: var
+      real(real64), contiguous, dimension(:, :, :), intent(IN) :: var
       character(len=*), intent(in) :: varname
       integer, intent(in), optional :: opt_mode
       type(d2d_io_family), target, intent(in), optional :: opt_family
@@ -358,7 +358,7 @@ contains
                         opt_family=opt_family, &
                         opt_io=opt_io, &
                         opt_dirname=opt_dirname, &
-                        freal=real(var, kind=kind(0._real32)))
+                        freal=real(var, kind=real32))
       else
          call write_one(ipencil, varname, decomp, &
                         opt_mode=opt_mode, &
@@ -378,7 +378,7 @@ contains
       implicit none
 
       integer, intent(IN) :: ipencil
-      complex(kind(0._real64)), contiguous, dimension(:, :, :), intent(IN) :: var
+      complex(real64), contiguous, dimension(:, :, :), intent(IN) :: var
       character(len=*), intent(in) :: varname
       integer, intent(in), optional :: opt_mode
       type(d2d_io_family), target, intent(in), optional :: opt_family
@@ -410,7 +410,7 @@ contains
                         opt_family=opt_family, &
                         opt_io=opt_io, &
                         opt_dirname=opt_dirname, &
-                        fcplx=cmplx(var, kind=kind(0._real32)))
+                        fcplx=cmplx(var, kind=real32))
       else
          call write_one(ipencil, varname, decomp, &
                         opt_mode=opt_mode, &
@@ -433,7 +433,7 @@ contains
       implicit none
 
       integer, intent(IN) :: ipencil
-      real(kind(0._real32)), contiguous, dimension(:, :, :), intent(out) :: var
+      real(real32), contiguous, dimension(:, :, :), intent(out) :: var
       character(len=*), intent(in) :: varname
       type(d2d_io_family), target, intent(in), optional :: opt_family
       type(d2d_io), intent(inout), optional :: opt_io
@@ -469,7 +469,7 @@ contains
       implicit none
 
       integer, intent(IN) :: ipencil
-      complex(kind(0._real32)), contiguous, dimension(:, :, :), intent(out) :: var
+      complex(real32), contiguous, dimension(:, :, :), intent(out) :: var
       character(len=*), intent(in) :: varname
       type(d2d_io_family), target, intent(in), optional :: opt_family
       type(d2d_io), intent(inout), optional :: opt_io
@@ -505,7 +505,7 @@ contains
[…]
