### 2023-09-22 ce1aeab
Exit directly when trying to initialize IO family without the ADIOS2 flag
