### 2023-09-22 192ec49
Update the io_read test for the new IO design
