### 2023-09-23 e72d3f7
Use default_opt_reduce_prec when a variable is registered
