### 2023-09-26 47ec7da
Adapt io_bench to the new IO design
