+++ examples/io_test/io_bench.f90
@@ -1,11 +1,12 @@
 !! SPDX-License-Identifier: BSD-3-Clause
 program io_bench
 
+   use decomp_2d
    use decomp_2d_constants
    use decomp_2d_io
+   use decomp_2d_io_object
    use decomp_2d_mpi
    use decomp_2d_testing
-   use decomp_2d
    use MPI
 #if defined(_GPU)
    use cudafor
@@ -21,6 +22,7 @@ program io_bench
    integer :: nranks_tot
 
    real(mytype), allocatable, dimension(:, :, :) :: u1
+   type(d2d_io) :: io
 
    double precision :: t1, t2
    integer :: ierror
@@ -40,11 +42,20 @@ program io_bench
 
    call decomp_2d_testing_log()
 
+   call decomp_2d_io_init()
+   call decomp_2d_io_register_var3d("io.dat", 1, real_type)
+
    call alloc_x(u1, .true.)
    call random_number(u1)
 
    t1 = MPI_WTIME()
-   call decomp_2d_write_one(1, u1, 'io.dat')
+#ifdef ADIOS2
+   call io%open_start("./", decomp_2d_write_mode)
+#endif
+   call decomp_2d_write_one(1, u1, 'io.dat', opt_dirname="./", opt_io=io)
+#ifdef ADIOS2
+   call io%end_close
+#endif
    t2 = MPI_WTIME()
 
    if (nrank == 0) then
@@ -55,6 +66,7 @@ program io_bench
    end if
 
    deallocate (u1)
+   call decomp_2d_io_fin
    call decomp_2d_finalize
    call MPI_FINALIZE(ierror)
 
