### 2023-09-26 2e3ccea
Fix MPI IO in case of ADIOS2 build
