+++ src/io.f90
@@ -47,8 +47,8 @@ module decomp_2d_io
              decomp_2d_write_subdomain, &
              decomp_2d_write_outflow, decomp_2d_read_inflow, &
              decomp_2d_io_init, decomp_2d_io_fin, & ! XXX: initialise/finalise 2decomp&fft IO module
-             decomp_2d_register_variable, &
              decomp_2d_io_register_var3d, &
+             decomp_2d_io_register_var2d, &
              gen_iodir_name
 
    ! Generic interface to handle multiple data types
@@ -92,9 +92,10 @@ module decomp_2d_io
    end interface decomp_2d_read_scalar
 
    interface decomp_2d_write_plane
-      module procedure write_plane_3d_real
-      module procedure write_plane_3d_complex
-      !     module procedure write_plane_2d
+      module procedure write_plane_freal
+      module procedure write_plane_dreal
+      module procedure write_plane_fcplx
+      module procedure write_plane_dcplx
    end interface decomp_2d_write_plane
 
    interface decomp_2d_write_every
@@ -135,9 +136,39 @@ contains
       logical, intent(in), optional :: opt_reduce_prec
       type(decomp_info), intent(in), optional :: opt_decomp
 
-      call default_io_family%register_var3d(varname, ipencil, type, opt_reduce_prec, opt_decomp)
+      call default_io_family%register_var3d(varname, ipencil, type, &
+                                            opt_reduce_prec=opt_reduce_prec, &
+                                            opt_decomp=opt_decomp)
 
    end subroutine decomp_2d_io_register_var3d
+
+   !
+   ! High-level. Register planes for the default family of readers / writers.
+   !
+   !    See io_family.f90
+   !
+   subroutine decomp_2d_io_register_var2d(varname, &
+                                          ipencil, &
+                                          type, &
+                                          opt_reduce_prec, &
+                                          opt_decomp, &
+                                          opt_nplanes)
+
+      implicit none
+
+      character(len=*), intent(in) :: varname
+      integer, intent(in) :: ipencil ! (x-pencil=1; y-pencil=2; z-pencil=3)
+      integer, intent(in) :: type
+      logical, intent(in), optional :: opt_reduce_prec
+      type(decomp_info), intent(in), optional :: opt_decomp
+      integer, intent(in), optional :: opt_nplanes
+
+      call default_io_family%register_var2d(varname, ipencil, type, &
+                                            opt_reduce_prec=opt_reduce_prec, &
+                                            opt_decomp=opt_decomp, &
+                                            opt_nplanes=opt_nplanes)
+
+   end subroutine decomp_2d_io_register_var2d
 
    !
    ! Initialize the IO module
@@ -623,6 +654,221 @@ contains
    end subroutine read_one_dcplx
 
    !
+   ! High-level. Using MPI-IO / ADIOS2 to write planes to a file
+   !
+   subroutine write_plane_freal(ipencil, var, varname, opt_nplanes, opt_mode, opt_family, &
+                                opt_io, opt_dirname, opt_reduce_prec, opt_decomp)
+
+      implicit none
+
+      integer, intent(IN) :: ipencil
+      real(real32), contiguous, dimension(:, :, :), intent(IN) :: var
+      character(len=*), intent(in) :: varname
+      integer, intent(in), optional :: opt_nplanes
+      integer, intent(in), optional :: opt_mode
+      type(d2d_io_family), target, intent(in), optional :: opt_family
[…]
