### 2023-09-26 806afab
Update CI to the new IO design
