### 2023-09-26 6c661e0
Refactor code for planes IO
