### 2023-10-09 fa9a276
Revert "Minor refactoring in transpose subroutines"
