### 2023-10-09 #5813 @CeciliaCoelho
@diehlpk Finished my review. Just some minor changes and for me it's good to go!

******************REVIEW****************************************
This software is well motivated and much effort has been put into it as it can be seen from the well built repository. The abstract and statement of need are very well written and convincing. However, there are some minor problems with the paper and repository that I would like the authors to address. In addition there is one major problem (in my opinion) that must be addressed prior publication as to ensure a more easy introduction to the library for beginners that is the addition of actual practical examples to the documentation, exemplifying the usage of the library in a more beginner-friendly way.

### Minor
1-text in line 53 ends halfway so text in line 54 should be fitted there.

2-There is a Contribute.md but it is not mentioned in the README. Could you add a small mention for people to go check it out if you don't want to merge it to the README? I undestand the authors might have chosen for it to be a separate file since the REAME is pretty long already.

### Major
1-it would be a great addition to have 1 or 2 practical examples in the paper document with the description of the problems and the code to use the 2DECOMP&FFT. This would be a nice addition to the README in the repository too in the "Testing and examples" section.
*************************************************
