+++ src/transpose_y_to_z.f90
@@ -39,7 +39,11 @@
         dst = src
 #endif
      else
-        call transpose_y_to_z_real(src, dst, decomp)
+#if defined(_GPU)
+        call transpose_y_to_z_real(src, dst, decomp, work1_r_d, work2_r_d)
+#else
+        call transpose_y_to_z_real(src, dst, decomp, work1_r, work2_r)
+#endif
      end if
 
 #ifdef PROFILER
@@ -48,13 +52,17 @@
 
   end subroutine transpose_y_to_z_real_long
 
-  subroutine transpose_y_to_z_real(src, dst, decomp)
+  subroutine transpose_y_to_z_real(src, dst, decomp, wk1, wk2)
 
      implicit none
 
      real(mytype), dimension(:, :, :), intent(IN) :: src
      real(mytype), dimension(:, :, :), intent(OUT) :: dst
      TYPE(DECOMP_INFO), intent(IN) :: decomp
+     real(mytype), dimension(:), intent(out) :: wk1, wk2
+#if defined(_GPU)
+     attributes(device) :: wk1, wk2
+#endif
 
 #if defined(_GPU)
      integer :: istat
@@ -71,48 +79,42 @@
      d3 = SIZE(dst, 3)
 
      ! rearrange source array as send buffer
-#if defined(_GPU)
-     call mem_split_yz_real(src, s1, s2, s3, work1_r_d, dims(2), &
+     call mem_split_yz_real(src, s1, s2, s3, wk1, dims(2), &
                             decomp%y2dist, decomp)
-#else
-     call mem_split_yz_real(src, s1, s2, s3, work1_r, dims(2), &
-                            decomp%y2dist, decomp)
-#endif
 
      ! define receive buffer
 #ifdef EVEN
      if (decomp%even) then
-        call MPI_ALLTOALL(work1_r, decomp%y2count, &
-                          real_type, dst, decomp%z2count, &
-                          real_type, DECOMP_2D_COMM_ROW, ierror)
-        if (ierror /= 0) call decomp_2d_abort(__FILE__, __LINE__, ierror, "MPI_ALLTOALLV")
+        call MPI_ALLTOALL(wk1, decomp%y2count, real_type, &
+                          dst, decomp%z2count, real_type, &
+                          DECOMP_2D_COMM_ROW, ierror)
      else
-        call MPI_ALLTOALL(work1_r, decomp%y2count, &
-                          real_type, work2_r, decomp%z2count, &
-                          real_type, DECOMP_2D_COMM_ROW, ierror)
-        if (ierror /= 0) call decomp_2d_abort(__FILE__, __LINE__, ierror, "MPI_ALLTOALL")
+        call MPI_ALLTOALL(wk1, decomp%y2count, real_type, &
+                          wk2, decomp%z2count, real_type, &
+                          DECOMP_2D_COMM_ROW, ierror)
      end if
+     if (ierror /= 0) call decomp_2d_abort(__FILE__, __LINE__, ierror, "MPI_ALLTOALL")
 #else
 
 #if defined(_GPU)
 #if defined(_NCCL)
-     call decomp_2d_nccl_send_recv_row(work2_r_d, &
-                                       work1_r_d, &
+     call decomp_2d_nccl_send_recv_row(wk2, &
+                                       wk1, &
                                        decomp%y2disp, &
                                        decomp%y2cnts, &
                                        decomp%z2disp, &
                                        decomp%z2cnts, &
                                        dims(2))
 #else
-     call MPI_ALLTOALLV(work1_r_d, decomp%y2cnts, decomp%y2disp, &
[…]
