### 2023-10-12 8db8472
Avoid unused variables
