+++ src/transpose_z_to_y.f90
@@ -128,6 +128,8 @@
      if (ierror /= 0) call decomp_2d_abort(__FILE__, __LINE__, ierror, "MPI_ALLTOALLV")
 #endif
 #else
+     associate (wk => wk1)
+     end associate
      call MPI_ALLTOALLV(src, decomp%z2cnts, decomp%z2disp, real_type, &
                         wk2, decomp%y2cnts, decomp%y2disp, real_type, &
                         DECOMP_2D_COMM_ROW, ierror)
@@ -267,6 +269,8 @@
      if (ierror /= 0) call decomp_2d_abort(__FILE__, __LINE__, ierror, "MPI_ALLTOALLV")
 #endif
 #else
+     associate (wk => wk1)
+     end associate
      call MPI_ALLTOALLV(src, decomp%z2cnts, decomp%z2disp, complex_type, &
                         wk2, decomp%y2cnts, decomp%y2disp, complex_type, &
                         DECOMP_2D_COMM_ROW, ierror)
