+++ src/transpose_y_to_z.f90
@@ -250,7 +250,7 @@
                                        decomp_buf_size)
 #else
      call MPI_ALLTOALLV(wk1, decomp%y2cnts, decomp%y2disp, complex_type, &
-                        wk2, decomp%z2cnts, decomp%z2disp, &complex_type, &
+                        wk2, decomp%z2cnts, decomp%z2disp, complex_type, &
                         DECOMP_2D_COMM_ROW, ierror)
      if (ierror /= 0) call decomp_2d_abort(__FILE__, __LINE__, ierror, "MPI_ALLTOALLV")
 #endif
