### 2023-10-12 e55e298
Fuse CPU and GPU memory buffers
