+++ src/decomp_2d_cumpi.f90
@@ -11,9 +11,9 @@ module decomp_2d_cumpi
 
    private        ! Make everything private unless declared public
 
-   ! Device working arrays
-   real(mytype), allocatable, dimension(:), device, public:: work1_r_d, work2_r_d
-   complex(mytype), allocatable, dimension(:), device, public :: work1_c_d, work2_c_d
+   ! Real/complex pointers to GPU buffers
+   real(mytype), dimension(:), device, public, pointer :: work1_r_d, work2_r_d
+   complex(mytype), dimension(:), device, public, pointer :: work1_c_d, work2_c_d
 
    public :: decomp_2d_cumpi_init, &
              decomp_2d_cumpi_fin
@@ -22,41 +22,21 @@ contains
    !
    ! init of the arrays
    !
-   subroutine decomp_2d_cumpi_init(buf_size)
+   subroutine decomp_2d_cumpi_init(buf_size, wk1, wk2)
 
       implicit none
 
       integer, intent(in) :: buf_size
-      integer :: status, errorcode
+      real(mytype), target, dimension(:), device :: wk1, wk2
 
-      if (allocated(work1_r_d)) deallocate (work1_r_d)
-      if (allocated(work2_r_d)) deallocate (work2_r_d)
-      if (allocated(work1_c_d)) deallocate (work1_c_d)
-      if (allocated(work2_c_d)) deallocate (work2_c_d)
-      allocate (work1_r_d(buf_size), STAT=status)
-      if (status /= 0) then
-         errorcode = 2
-         call decomp_2d_abort(__FILE__, __LINE__, errorcode, &
-                              'Out of memory when allocating 2DECOMP workspace')
-      end if
-      allocate (work1_c_d(buf_size), STAT=status)
-      if (status /= 0) then
-         errorcode = 2
-         call decomp_2d_abort(__FILE__, __LINE__, errorcode, &
-                              'Out of memory when allocating 2DECOMP workspace')
-      end if
-      allocate (work2_r_d(buf_size), STAT=status)
-      if (status /= 0) then
-         errorcode = 2
-         call decomp_2d_abort(__FILE__, __LINE__, errorcode, &
-                              'Out of memory when allocating 2DECOMP workspace')
-      end if
-      allocate (work2_c_d(buf_size), STAT=status)
-      if (status /= 0) then
-         errorcode = 2
-         call decomp_2d_abort(__FILE__, __LINE__, errorcode, &
-                              'Out of memory when allocating 2DECOMP workspace')
-      end if
+      if (associated(work1_r_d)) nullify (work1_r_d)
+      if (associated(work2_r_d)) nullify (work2_r_d)
+      if (associated(work1_c_d)) nullify (work1_c_d)
+      if (associated(work2_c_d)) nullify (work2_c_d)
+      call c_f_pointer(c_loc(wk1), work1_r_d, [buf_size])
+      call c_f_pointer(c_loc(wk2), work2_r_d, [buf_size])
+      call c_f_pointer(c_loc(wk1), work1_c_d, [buf_size])                                            
+      call c_f_pointer(c_loc(wk2), work2_c_d, [buf_size])
 
    end subroutine decomp_2d_cumpi_init
    !
@@ -66,7 +46,10 @@ contains
 
       implicit none
 
-      deallocate (work1_r_d, work2_r_d, work1_c_d, work2_c_d)
+      if (associated(work1_r_d)) nullify (work1_r_d)
+      if (associated(work2_r_d)) nullify (work2_r_d)                                                 
+      if (associated(work1_c_d)) nullify (work1_c_d)
+      if (associated(work2_c_d)) nullify (work2_c_d)
 
    end subroutine decomp_2d_cumpi_fin
 
