### 2023-10-15 c3c5302
Use the dedicated allocator whenever possible
