### 2023-10-15 8f260ae
CI + gpu : add a test for the cufft backend with inplace FFT
