### 2023-10-15 63c051a
Revert "CI + gpu : add a test for the cufft backend with inplace FFT"
This reverts commit 8f260aea8a60daecc7d087e95ede1f61d81e56ef.
