### 2023-10-16 3ce989d
Update the visu example to support only real and use GPU
