+++ examples/io_test/io_visu.f90
@@ -17,11 +17,7 @@ program visu
   integer :: nx, ny, nz
   integer :: p_row = 0, p_col = 0
 
-#ifdef COMPLEX_TEST
-  complex(mytype), allocatable, dimension(:, :, :) :: u1, u2, u3
-#else
   real(mytype), allocatable, dimension(:, :, :) :: u1, u2, u3
-#endif
 
   integer, parameter :: output2D = 0 ! Which plane to write in 2D (0 for 3D)
   character(len=*), parameter :: io_name = "visu-io"
@@ -101,18 +97,9 @@ contains
 
     call decomp_2d_io_init()
     call decomp_2d_init_io(io_name)
-#ifndef COMPLEX_TEST
     call decomp_2d_register_variable(io_name, "u1.dat", 1, 0, output2D, mytype)
     call decomp_2d_register_variable(io_name, "u2.dat", 2, 0, output2D, mytype)
     call decomp_2d_register_variable(io_name, "u3.dat", 3, 0, output2D, mytype)
-#else
-    call decomp_2d_register_variable(io_name, "u1_re.dat", 1, 0, output2D, mytype)
-    call decomp_2d_register_variable(io_name, "u2_re.dat", 2, 0, output2D, mytype)
-    call decomp_2d_register_variable(io_name, "u3_re.dat", 3, 0, output2D, mytype)
-    call decomp_2d_register_variable(io_name, "u1_im.dat", 1, 0, output2D, mytype)
-    call decomp_2d_register_variable(io_name, "u2_im.dat", 2, 0, output2D, mytype)
-    call decomp_2d_register_variable(io_name, "u3_im.dat", 3, 0, output2D, mytype)
-#endif
 
   end subroutine init_example
 
@@ -122,39 +109,40 @@ contains
     call alloc_y(u2)
     call alloc_z(u3)
     
+    ! distribute the data
+    !$acc data copy(u1,u2,u3)
+
+    !$acc parallel loop default(present)
     do k = 1, xsize(3)
        do j = 1, xsize(2)
           do i = 1, xsize(1)
-#ifndef COMPLEX_TEST
-             u1(i, j, k) = real(nrank, mytype)
-#else
-             u1(i, j, k) = cmplx(real(nrank, mytype), -real(nrank, mytype))
-#endif
+             u1(i, j, k) = real(100+nrank, mytype)
           end do
        end do
     end do
+    !$acc end loop
+    !$acc parallel loop default(present)
     do k = 1, ysize(3)
        do j = 1, ysize(2)
           do i = 1, ysize(1)
-#ifndef COMPLEX_TEST
-             u2(i, j, k) = real(nrank, mytype)
-#else
-             u2(i, j, k) = cmplx(real(nrank, mytype), -real(nrank, mytype))
-#endif
+             u2(i, j, k) = real(100+nrank, mytype)
           end do
        end do
     end do
+    !$acc end loop
+    !$acc parallel loop default(present)
     do k = 1, zsize(3)
        do j = 1, zsize(2)
           do i = 1, zsize(1)
-#ifndef COMPLEX_TEST
-             u3(i, j, k) = real(nrank, mytype)
-#else
-             u3(i, j, k) = cmplx(real(nrank, mytype), -real(nrank, mytype))
-#endif
+             u3(i, j, k) = real(100+nrank, mytype)
           end do
        end do
     end do
+    !$acc end loop
+    !$acc update self (u1)
+    !$acc update self (u2)
+    !$acc update self (u3)
+    !$acc end data
     
   end subroutine init_data
 
[…]
