+++ examples/CMakeLists.txt
@@ -14,4 +14,4 @@ add_subdirectory(halo_test)
 add_subdirectory(io_test)
 
 # Set real/complex tests
-set(COMPLEX_TESTS "OFF" CACHE STRING "Enables complex numbers for tests that support it")
+#set(COMPLEX_TESTS "OFF" CACHE STRING "Enables complex numbers for tests that support it")
