+++ examples/io_test/io_visu.f90
@@ -4,277 +4,277 @@
 
 program visu
 
-  use mpi
+   use mpi
 
-  use decomp_2d
-  use decomp_2d_mpi
-  use decomp_2d_constants
-  use decomp_2d_io
+   use decomp_2d
+   use decomp_2d_mpi
+   use decomp_2d_constants
+   use decomp_2d_io
 
-  implicit none
+   implicit none
 
-  integer, parameter :: nx_base = 17, ny_base = 13, nz_base = 11
-  integer :: nx, ny, nz
-  integer :: p_row = 0, p_col = 0
+   integer, parameter :: nx_base = 17, ny_base = 13, nz_base = 11
+   integer :: nx, ny, nz
+   integer :: p_row = 0, p_col = 0
 
-  real(mytype), allocatable, dimension(:, :, :) :: u1, u2, u3
+   real(mytype), allocatable, dimension(:, :, :) :: u1, u2, u3
 
-  integer, parameter :: output2D = 0 ! Which plane to write in 2D (0 for 3D)
-  character(len=*), parameter :: io_name = "visu-io"
+   integer, parameter :: output2D = 0 ! Which plane to write in 2D (0 for 3D)
+   character(len=*), parameter :: io_name = "visu-io"
 #ifndef ADIOS2
-  logical :: dir_exists
+   logical :: dir_exists
 #endif
 
-  integer :: nargin, arg, FNLength, status, DecInd
+   integer :: nargin, arg, FNLength, status, DecInd
 
-  integer :: i, j, k
+   integer :: i, j, k
 
-  integer :: ierr
+   integer :: ierr
 
-  call init_example()
-  call init_data()
+   call init_example()
+   call init_data()
 
-  call write_data()
-  call write_visu()
-  
-  call fin()
-  
+   call write_data()
+   call write_visu()
+
+   call fin()
+
 contains
 
-  subroutine init_example()
+   subroutine init_example()
 
-    integer :: resize_domain
-    integer :: nranks_tot
-    character(len=80) :: InputFN
+      integer :: resize_domain
+      integer :: nranks_tot
+      character(len=80) :: InputFN
 
-    call MPI_INIT(ierr)
-    ! To resize the domain we need to know global number of ranks
-    ! This operation is also done as part of decomp_2d_init
-    call MPI_COMM_SIZE(MPI_COMM_WORLD, nranks_tot, ierr)
-    resize_domain = int(nranks_tot / 4) + 1
-    nx = nx_base * resize_domain
-    ny = ny_base * resize_domain
-    nz = nz_base * resize_domain
-    ! Now we can check if user put some inputs
-    ! Handle input file like a boss -- GD
-    nargin = command_argument_count()
-    if ((nargin == 0) .or. (nargin == 2) .or. (nargin == 5)) then
-       do arg = 1, nargin
-          call get_command_argument(arg, InputFN, FNLength, status)
-          read (InputFN, *, iostat=status) DecInd
-          if (arg == 1) then
-             p_row = DecInd
-          elseif (arg == 2) then
-             p_col = DecInd
-          elseif (arg == 3) then
-             nx = DecInd
-          elseif (arg == 4) then
-             ny = DecInd
-          elseif (arg == 5) then
-             nz = DecInd
-          end if
-       end do
-    else
-       ! nrank not yet computed we need to avoid write
-       ! for every rank
-       call MPI_COMM_RANK(MPI_COMM_WORLD, nrank, ierr)
-       if (nrank == 0) then
-          print *, "This Test takes no inputs or 2 inputs as"
-          print *, "  1) p_row (default=0)"
[…]
