### 2023-10-16 f395b8e
Bug correction for GPU: add scalar for end index loops
