### 2023-10-21 4be2925
Add profiling to the FFTW backend (#274)
