+++ src/fft_fftw3.f90
@@ -474,6 +474,10 @@ module decomp_2d_fft
       complex(mytype), allocatable, dimension(:, :, :) :: wk1
 #endif
 
+#ifdef PROFILER
+      if (decomp_profiler_fft) call decomp_profiler_start("fft_c2c")
+#endif
+
       if (format == PHYSICAL_IN_X .AND. isign == DECOMP_2D_FFT_FORWARD .OR. &
           format == PHYSICAL_IN_Z .AND. isign == DECOMP_2D_FFT_BACKWARD) then
 
@@ -557,6 +561,10 @@ module decomp_2d_fft
       deallocate (wk1)
 #endif
 
+#ifdef PROFILER
+      if (decomp_profiler_fft) call decomp_profiler_end("fft_c2c")
+#endif
+
       return
    end subroutine fft_3d_c2c
 
@@ -569,6 +577,10 @@ module decomp_2d_fft
 
       real(mytype), dimension(:, :, :), intent(INOUT) :: in_r
       complex(mytype), dimension(:, :, :), intent(OUT) :: out_c
+
+#ifdef PROFILER
+      if (decomp_profiler_fft) call decomp_profiler_start("fft_r2c")
+#endif
 
       if (format == PHYSICAL_IN_X) then
 
@@ -613,6 +625,10 @@ module decomp_2d_fft
 
       end if
 
+#ifdef PROFILER
+      if (decomp_profiler_fft) call decomp_profiler_end("fft_r2c")
+#endif
+
       return
    end subroutine fft_3d_r2c
 
@@ -628,6 +644,10 @@ module decomp_2d_fft
 
 #ifndef OVERWRITE
       complex(mytype), allocatable, dimension(:, :, :) :: wk1
+#endif
+
+#ifdef PROFILER
+      if (decomp_profiler_fft) call decomp_profiler_start("fft_c2r")
 #endif
 
       if (format == PHYSICAL_IN_X) then
@@ -700,6 +720,10 @@ module decomp_2d_fft
 
 #ifndef OVERWRITE
       deallocate (wk1)
+#endif
+
+#ifdef PROFILER
+      if (decomp_profiler_fft) call decomp_profiler_end("fft_c2r")
 #endif
 
       return
