### 2023-10-22 38718cf
Use the dedicated allocator whenever possible (#272)
