+++ src/fft_mkl.f90
@@ -359,7 +359,7 @@ module decomp_2d_fft
 #ifdef OVERWRITE
          status = wrapper_c2c_inplace(c2c_x, in, isign)
 #else
-         allocate (wk1(ph%xsz(1), ph%xsz(2), ph%xsz(3)))
+         call alloc_x(wk1, ph)
          status = wrapper_c2c(c2c_x, in, wk1, isign)
 #endif
          if (status /= 0) call decomp_2d_abort(__FILE__, __LINE__, status, "wrapper_c2c")
@@ -378,7 +378,7 @@ module decomp_2d_fft
             if (status /= 0) call decomp_2d_abort(__FILE__, __LINE__, status, "wrapper_c2c")
          end do
 #else
-         allocate (wk2b(ph%ysz(1), ph%ysz(2), ph%ysz(3)))
+         call alloc_y(wk2b, ph)
          do k = 1, ph%xsz(3) ! one Z-plane at a time
             !          if (isign==DECOMP_2D_FFT_FORWARD) then
             !             status = DftiComputeForward(c2c_y, wk2(:,1,k), wk2b(:,1,k))
@@ -394,7 +394,7 @@ module decomp_2d_fft
 #ifdef OVERWRITE
          call transpose_y_to_z(wk2_c2c, out, ph)
 #else
-         allocate (wk3(ph%zsz(1), ph%zsz(2), ph%zsz(3)))
+         call alloc_z(wk3, ph)
          call transpose_y_to_z(wk2b, wk3, ph)
 #endif
 
@@ -424,7 +424,7 @@ module decomp_2d_fft
 #ifdef OVERWRITE
          status = wrapper_c2c_inplace(c2c_z, in, isign)
 #else
-         allocate (wk1(ph%zsz(1), ph%zsz(2), ph%zsz(3)))
+         call alloc_z(wk1, ph)
          status = wrapper_c2c(c2c_z, in, wk1, isign)
 #endif
          if (status /= 0) call decomp_2d_abort(__FILE__, __LINE__, status, "wrapper_c2c")
@@ -443,7 +443,7 @@ module decomp_2d_fft
             if (status /= 0) call decomp_2d_abort(__FILE__, __LINE__, status, "wrapper_c2c")
          end do
 #else
-         allocate (wk2b(ph%ysz(1), ph%ysz(2), ph%ysz(3)))
+         call alloc_y(wk2b, ph)
          do k = 1, ph%xsz(3) ! one Z-plane at a time
             !          if (isign==DECOMP_2D_FFT_FORWARD) then
             !             status = DftiComputeForward(c2c_y, wk2(:,1,k), wk2b(:,1,k))
@@ -459,7 +459,7 @@ module decomp_2d_fft
 #ifdef OVERWRITE
          call transpose_y_to_x(wk2_c2c, out, ph)
 #else
-         allocate (wk3(ph%xsz(1), ph%xsz(2), ph%xsz(3)))
+         call alloc_x(wk3, ph)
          call transpose_y_to_x(wk2b, wk3, ph)
 #endif
 
@@ -528,7 +528,7 @@ module decomp_2d_fft
             if (status /= 0) call decomp_2d_abort(__FILE__, __LINE__, status, "wrapper_c2c")
          end do
 #else
-         allocate (wk2b(sp%ysz(1), sp%ysz(2), sp%ysz(3)))
+         call alloc_y(wk2b, sp)
          do k = 1, sp%ysz(3)
             !          status = DftiComputeForward(c2c_y2, wk2(:,1,k), wk2b(:,1,k))
[…]
