+++ examples/fft_physical_z/README.md
@@ -1,5 +1,9 @@
-fft_c2c_z, fft_r2c_z
-------
+# Test FFT for Z pencil decomposition 
+
+List of the tests:
+- [fft_c2c_z](fft_c2c_z.f90): Test Complex to Complex FFT transform; 
+- [fft_r2c_x](fft_r2c_z.f90): Test Real to Complex FFT transform;
+
 
 These programs can be used to test the fft trasform using Z as starting physical dimension. 
 Both c2c (fft_c2c_z) and r2c/c2r (fft_r2c_z) transforms are tested.
@@ -8,14 +12,16 @@ after a forward and a backward transform and proper normalisation.
 The test automatically resize the problem depending on the number of MPI processes in use
 
 What to input: The program takes max 6 inputs as : 
-               1) p_row [optional]
-	       2) p_col [optional] 
-	       3) nx    [optional]
-	       4) ny    [optional]
-	       5) nz    [optional]
-	       6) nt    [optional]
-	       In case the decomposition is imposed both (1) and (2) are necessary. 
-	       If also the resolution is imposed (1-5) are necessary
+
+1. p_row [optional]
+1. p_col [optional] 
+1. nx    [optional]
+1. ny    [optional]
+1. nz    [optional]
+1. nt    [optional]
+
+In case the decomposition is imposed both (1) and (2) are necessary. 
+If also the resolution is imposed (1-5) are necessary
 
 What to expect:
 - The timing results 
