+++ HOWTO.md
@@ -1,0 +1,9 @@
+# How To use 2DECOMP&FFT
+
+The 2D Pencil Decomposition API is defined in three Fortran module which should be used by applications as:
+
+::
+
+  use decomp_2d_constants
+  use decomp_2d_mpi
+  use decomp_2d
