### 2023-10-24 7dfd3ae
Minor update in HOWTO.md
