+++ HOWTO.md
@@ -1,14 +1,14 @@
 # How To use 2DECOMP&FFT
 Full informations on how to use the 2DECOMP&FFT library together with some related theory are available [here](https://2decomp-fft.github.io/).
 
-The 2D Pencil Decomposition API is defined in three Fortran module which should be used by applications as:
+The 2D Pencil Decomposition API is defined in three Fortran modules which should be used by applications as:
 ```
   use decomp_2d_constants
   use decomp_2d_mpi
   use decomp_2d
 ```
 where ``use decomp_2d_constants`` defines all the parameters, ``use decomp_2d_mpi`` introduces all the MPI 
-related interfaces and ``use decomp_2d`` cointains the main decomposition and transposition APIs. 
+related interfaces and ``use decomp_2d`` cointains the main decomposition and transposition APIs. The library is initialized using
 ```
   call decomp_2d_init(nx, ny, nz, p_row, p_col)
 ```
@@ -25,23 +25,23 @@ Correspondingly, the library provides 4 communication subroutines:
   call transpose_z_to_y(var_in,var_out)
   call transpose_y_to_x(var_in,var_out)
 ```
-The input array ``var_in`` and ``var_output`` array out should have been defined
+The input array ``var_in`` and output array ``var_out`` are defined by the code using the library
 and contain distributed data for the correct pencil orientations.
 
 Note that the library is written using Fortran's generic interface so different data types are supported
-without user intervention. That means in and out above can be either real arrays or complex arrays,
+without user intervention. Arrays above can be either real arrays or complex ones,
 the latter being useful for FFT-type of applications.
-Finally, before exit, applications should clean up the memory by:
+Finally, before exit, applications should clean up the memory using
 ```
   call decomp_2d_finalize
 ```
 Detailed information about the decomposition API are available [here](https://2decomp-fft.github.io/pages/api_domain.html) 
 ### Use of the FFT module
-To use the FFT programming interface, first of all, one additional Fortran module has to be used:
+To use the FFT programming interface, first of all, one additional Fortran module is needed:
 ```
   use decomp_2d_fft
 ```
-Then one needs to initialise the FFT interface by:
+Then one needs to initialise the FFT interface using
 ```
   call decomp_2d_fft_init(pencil, n1, n2, n3)
 ```
