### 2023-10-24 aa4e467
Typo in INSTALL.md
