+++ README.md
@@ -3,7 +3,7 @@
 This README contains basic instructions for building and installing the 2DECOMP&FFT library, more
 detailed instructions about installation and linking to the library within an external project 
 can be found in the [install section](INSTALL.md).
-For how to use the library within your projec please refers to [HOWTO.md](HOWTO.md) and at the [examples](examples/README.md) 
+For how to use the library within your project please refers to [HOWTO.md](HOWTO.md) and at the [examples](examples/README.md) 
 provided. 
 
 ## Building
