### 2023-10-25 756ccaf
Minor update in HOWTO.md
