+++ HOWTO.md
@@ -53,7 +53,7 @@ For complex-to-complex (c2c) FFTs, the user interface is:
 ```
 where ``direction`` can be either ``DECOMP_2D_FFT_FORWARD == -1`` for forward transforms, 
 or ``DECOMP_2D_FFT_BACKWARD == 1`` for backward transforms.
-The input array ``input`` and ``output`` array out are both complex
+The input array (``input``) and the output one (``output``) are both complex
 and have to be either a X-pencil/Z-pencil combination or vice-versa.
 The interface for the the real-to-complex and complex-to-real transform is
 ```
