+++ examples/README.md
@@ -1,15 +1,10 @@
 # Examples
 
-* [Initialization test](init_test)         - to test the initialisation of the DECOMP2D&FFT library
-
-* [Test decomposition](test2d)             - various tests for the 2D pencil decomposition module and timing 
-
-* [Test FFT from X pencil](fft_physical_x) - various tests for the FFT starting from the ``X`` direction 
-
-* [Test FFT from Z pencil](fft_physical_z) - various tests for the FFT starting from the ``Z`` direction 
-
-* [Test HALO exchange](halo_test)          - to test the halo-cell exchange code
-
-* [Test IO features](io_test)              - various tests for the IO module
+- [Initialization test](init_test)         - to test the initialisation of the DECOMP2D&FFT library
+- [Test decomposition](test2d)             - various tests for the 2D pencil decomposition module and timing 
+- [Test FFT from X pencil](fft_physical_x) - various tests for the FFT starting from the ``X`` direction 
+- [Test FFT from Z pencil](fft_physical_z) - various tests for the FFT starting from the ``Z`` direction 
+- [Test HALO exchange](halo_test)          - to test the halo-cell exchange capability
+- [Test IO features](io_test)              - various tests for the IO module
 
 Refer to the README files for each example for details.
