### 2023-10-25 c61dc4b
Minor edits in HOWTO.md
Minor rewording, and highlighting that load imbalance will occur for arbitrary grid/decomposition combinations
