### 2023-10-25 aef63b5
Update HOWTO.md
Users shouldn't use +/-1 directly in calls to FFT direction
