### 2023-10-26 #5813 @p-costa
It was a pleasure to review 2DECOMP&FFT, and I am sorry for taking longer than I expected to provide feedback. Apart from the paper, 2DECOMP&FFT ticks all the boxes above, except for the missing examples of the API in use as mentioned in the other review. These could be quite useful to new users. I remember good examples from the 2012 2DECOMP&FFT website on how to compute vorticity which I found pretty insightful. These could be recycled in these docs and updated for the new API (?). I think CMAKE will be a barrier to some Fortran users, but at least some guidelines for setting up a Makefile are also provided, so that's great.

 Regarding the paper, I have a few comments for your consideration:

   - I think it is risky to have in the title "The new version of ...", because it is not timeless. Would this title be accurate ten years from now? I understand the reasoning behind the current title, but I suggest you think of something that does not deprecate with time.
   - I see some repetitions in the summary and the first sentences of the statement of need.
   - "It is designed for ... and spatially implicit numerical algorithms" I understand that this is what motivated 2decomp, but written this way it sounds like explicit finite-difference codes won't benefit from the library.
   - I found some typos, I suggest running the paper through a grammar and spell checker. L11: mesh -> meshes; L28-L29: the enumeration in bullet form misses the line break and semi-colon; L29:  there is a stop missing before "In"; L33: 3D tool -> tool to perform ... in three-dimensional domains (?); L31: "to perfor(m)"; L40: "has been design(ed)" L53: unintentional line break;
   - I found two papers under `paper.bib` that ended up not being cited in the built pdf: the cuDecomp paper (https://doi.org/10.1145/3539781.3539797), which is a new NVIDIA library inspired by 2DECOMP&FFT that provides optimized transpose functionality on NVIDIA GPUs (disclaimer: I co-author this paper); and the AFiD-GPU paper, which among other things adapted a previous version of the 2DECOMP&FFT library to operate on GPUs. I think these references are very relevant and should indeed be credited in the manuscript.
   - Finally, related to this, I browsed through the 2decomp source, and I think parts of the GPU implementation were inherited from the work done in the AFiD code? E.g.: https://github.com/PhysicsofFluids/AFiD_GPU_opensource/blob/master/transpose_x_to_y.F90#L563 vs https://github.com/2decomp-fft/2decomp-fft/blob/dev/src/transpose_x_to_y.f90#L326. I believe that this needs to be appropriately credited in the paper.
