### 2023-11-02 d35d0ec
Simplify the example io_bench
