### 2023-11-02 1d107a4
Fix example io_visu
