### 2023-11-02 0bd6de6
Update io_visu.f90
Use decomp_2d_testing
