+++ examples/io_test/io_visu.f90
@@ -10,6 +10,7 @@ program visu
    use decomp_2d_constants
    use decomp_2d_io
    use decomp_2d_mpi
+   use decomp_2d_testing
 
    implicit none
 
