### 2023-11-02 a8eea82
Use decomp_2d_testing in io_visu.f90 (#278)
* Update io_visu.f90

Use decomp_2d_testing

* Fix previous commit

Fix https://github.com/2decomp-fft/2decomp-fft/commit/0bd6de6f1a10647bf904702c89bf388b354655bf
