### 2023-11-03 c09a250
Adapt DEFAULT_OPT_REDUCE_PREC to the preprocessor variable SAVE_SINGLE
