### 2023-11-04 1e3ee70
Add a python script to generate the high-level code for writing one 3D file
