### 2023-11-04 7e473e5
Add a python script to generate the high-level code for reading one 3D file
