### 2023-11-04 23f7469
Add a python script to generate high-level code for writing planes
