+++ scripts/gen_io_write_plane.py
@@ -1,0 +1,222 @@
+#!/usr/bin/env python3
+
+# Name of the output file
+outfile = "write_plane.f90"
+
+# Number of variables
+nformat = 4
+# Suffix for the subroutine name
+ext = ["freal", "fcplx", "dreal", "dcplx"]
+
+
+#
+# Open the file
+#
+f = open(outfile, "w")
+
+#
+# Write the interface
+#
+f.write("\n")
+f.write("   interface decomp_2d_write_plane\n")
+for i in range(nformat):
+    f.write("      module procedure write_plane_"+ext[i]+"\n")
+
+f.write("   end interface decomp_2d_write_plane\n")
+f.write("\n")
+
+for i in range(nformat):
+    #
+    # Header
+    #
+    f.write("   subroutine write_plane_"+ext[i]+"(ipencil, var, varname, &\n")
+    f.write("                                opt_nplanes, opt_iplane, opt_mode, opt_family, &\n")
+    f.write("                                opt_io, opt_dirname, opt_reduce_prec, opt_decomp)\n")
+    f.write("\n")
+    f.write("      implicit none\n")
+    f.write("\n")
+    #
+    # Arguments
+    #
+    f.write("      integer, intent(IN) :: ipencil\n")
+    if (i==0):
+        f.write("      real(real32), contiguous, dimension(:, :, :), intent(IN) :: var\n")
+    elif (i==2):
+        f.write("      real(real64), contiguous, dimension(:, :, :), intent(IN) :: var\n")
+    elif (i==1):
+        f.write("      complex(real32), contiguous, dimension(:, :, :), intent(IN) :: var\n")
+    elif (i==3):
+        f.write("      complex(real64), contiguous, dimension(:, :, :), intent(IN) :: var\n")
+    f.write("      character(len=*), intent(in) :: varname\n")
+    f.write("      integer, intent(in), optional :: opt_nplanes, opt_iplane\n")
+    f.write("      integer, intent(in), optional :: opt_mode\n")
+    f.write("      type(d2d_io_family), target, intent(in), optional :: opt_family\n")
+    f.write("      type(d2d_io), intent(inout), optional :: opt_io\n")
+    f.write("      character(len=*), intent(in), optional :: opt_dirname\n")
+    f.write("      logical, intent(in), optional :: opt_reduce_prec\n")
+    f.write("      TYPE(DECOMP_INFO), target, intent(IN), optional :: opt_decomp\n")
+    f.write("\n")
+    #
+    # Local variables
+    #
+    if (i==2 or i==3):
+        f.write("      logical :: reduce\n")
+    if (i==0):
+        f.write("      real(real32), allocatable, dimension(:, :, :) :: var2d\n")
+    elif (i==1):
+        f.write("      complex(real32), allocatable, dimension(:, :, :) :: var2d\n")
+    elif (i==2):
+        f.write("      real(real64), allocatable, dimension(:, :, :) :: var2d\n")
+    elif (i==3):
+        f.write("      complex(real64), allocatable, dimension(:, :, :) :: var2d\n")
+    f.write("      TYPE(DECOMP_INFO), pointer :: decomp\n")
+    f.write("      integer :: nplanes, iplane\n")
+    f.write("\n")
+    #
+    # Deal with optional arguments
+    #
+    f.write("      if (present(opt_nplanes)) then\n")
+    f.write("         nplanes = opt_nplanes\n")
+    f.write("      else\n")
+    f.write("         nplanes = 1\n")
+    f.write("      end if\n")
+    f.write("\n")
+    f.write("      if (present(opt_iplane)) then\n")
+    f.write("         iplane = opt_iplane\n")
+    f.write("      else\n")
[…]
