+++ examples/io_test/io_tmp_test.f90
@@ -1,0 +1,388 @@
+!!! io_tmp_test.f90
+!!
+!! Tests writing from temporary arrays using the flush functionality, only relevant for ADIOS2.
+!!
+!! SPDX-License-Identifier: BSD-3-Clause
+program io_test
+
+   use mpi
+   use decomp_2d_constants
+   use decomp_2d_mpi
+   use decomp_2d
+   use decomp_2d_io
+#if defined(_GPU)
+   use cudafor
+   use openacc
+#endif
+
+   implicit none
+
+   integer, parameter :: nx_base = 17, ny_base = 13, nz_base = 11
+   integer :: nx, ny, nz
+   integer :: p_row = 0, p_col = 0
+   integer :: resize_domain
+   integer :: nranks_tot
+   integer :: nargin, arg, FNLength, status, DecInd
+   character(len=80) :: InputFN
+
+   integer :: ierr
+
+#ifdef COMPLEX_TEST
+   complex(mytype), allocatable, dimension(:, :, :) :: data1
+
+   complex(mytype), allocatable, dimension(:, :, :) :: u1, u2, u3
+   complex(mytype), allocatable, dimension(:, :, :) :: u1b, u2b, u3b
+   complex(mytype), allocatable, dimension(:, :, :) :: v1, v2, v3
+   complex(mytype), allocatable, dimension(:, :, :) :: v1b, v2b, v3b
+
+   complex(mytype), allocatable, dimension(:, :, :) :: tmp1, tmp2, tmp3
+#else
+   real(mytype), allocatable, dimension(:, :, :) :: data1
+
+   real(mytype), allocatable, dimension(:, :, :) :: u1, u2, u3
+   real(mytype), allocatable, dimension(:, :, :) :: u1b, u2b, u3b
+   real(mytype), allocatable, dimension(:, :, :) :: v1, v2, v3
+   real(mytype), allocatable, dimension(:, :, :) :: v1b, v2b, v3b
+
+   real(mytype), allocatable, dimension(:, :, :) :: tmp1, tmp2, tmp3
+#endif
+
+   real(mytype), parameter :: eps = 1.0E-7_mytype
+
+   character(len=*), parameter :: io_name = "test-io"
+   character(len=*), parameter :: io_restart = "restart-io"
+
+   integer :: i, j, k, m, ierror
+   integer :: xst1, xst2, xst3
+   integer :: xen1, xen2, xen3
+
+#ifndef ADIOS2
+   logical :: dir_exists
+#endif
+
+   integer, parameter :: output2D = 0 ! Which plane to write in 2D (0 for 3D)
+
+   call MPI_INIT(ierror)
+   ! To resize the domain we need to know global number of ranks
+   ! This operation is also done as part of decomp_2d_init
+   call MPI_COMM_SIZE(MPI_COMM_WORLD, nranks_tot, ierror)
+   resize_domain = int(nranks_tot / 4) + 1
+   nx = nx_base * resize_domain
+   ny = ny_base * resize_domain
+   nz = nz_base * resize_domain
+   ! Now we can check if user put some inputs
+   ! Handle input file like a boss -- GD
+   nargin = command_argument_count()
+   if ((nargin == 0) .or. (nargin == 2) .or. (nargin == 5)) then
+      do arg = 1, nargin
+         call get_command_argument(arg, InputFN, FNLength, status)
+         read (InputFN, *, iostat=status) DecInd
+         if (arg == 1) then
+            p_row = DecInd
+         elseif (arg == 2) then
+            p_col = DecInd
+         elseif (arg == 3) then
+            nx = DecInd
[…]
