### 2023-11-05 bac7c99
Add missing deallocations in io_tmp_test
