### 2023-11-05 de3eac4
Update paper.md acknowledgements
Add acknowledgement of eCSE funding
