### 2023-11-06 84c49b5
io_write_one : move profiling to the high-level subroutines
