+++ src/io.f90
@@ -278,6 +278,7 @@ contains
 
       implicit none
 
+      ! Arguments
       integer, intent(IN) :: ipencil
       real(real32), contiguous, dimension(:, :, :), intent(IN) :: var
       character(len=*), intent(in) :: varname
@@ -288,7 +289,12 @@ contains
       logical, intent(in), optional :: opt_reduce_prec
       TYPE(DECOMP_INFO), target, intent(IN), optional :: opt_decomp
 
+      ! Local variable(s)
       TYPE(DECOMP_INFO), pointer :: decomp
+
+#ifdef PROFILER
+      if (decomp_profiler_io) call decomp_profiler_start("io_write_one")
+#endif
 
       ! Use the provided decomp_info or the default one
       if (present(opt_decomp)) then
@@ -309,6 +315,10 @@ contains
       associate (p => opt_reduce_prec)
       end associate
 
+#ifdef PROFILER
+      if (decomp_profiler_io) call decomp_profiler_end("io_write_one")
+#endif
+
    end subroutine write_one_freal
    !
    subroutine write_one_fcplx(ipencil, var, varname, opt_mode, opt_family, &
@@ -316,6 +326,7 @@ contains
 
       implicit none
 
+      ! Arguments
       integer, intent(IN) :: ipencil
       complex(real32), contiguous, dimension(:, :, :), intent(IN) :: var
       character(len=*), intent(in) :: varname
@@ -326,7 +337,12 @@ contains
       logical, intent(in), optional :: opt_reduce_prec
       TYPE(DECOMP_INFO), target, intent(IN), optional :: opt_decomp
 
+      ! Local variable(s)
       TYPE(DECOMP_INFO), pointer :: decomp
+
+#ifdef PROFILER
+      if (decomp_profiler_io) call decomp_profiler_start("io_write_one")
+#endif
 
       ! Use the provided decomp_info or the default one
       if (present(opt_decomp)) then
@@ -347,6 +363,10 @@ contains
       associate (p => opt_reduce_prec)
       end associate
 
+#ifdef PROFILER
+      if (decomp_profiler_io) call decomp_profiler_end("io_write_one")
+#endif
+
    end subroutine write_one_fcplx
    !
    subroutine write_one_dreal(ipencil, var, varname, opt_mode, opt_family, &
@@ -354,6 +374,7 @@ contains
 
       implicit none
 
+      ! Arguments
       integer, intent(IN) :: ipencil
       real(real64), contiguous, dimension(:, :, :), intent(IN) :: var
       character(len=*), intent(in) :: varname
@@ -364,8 +385,13 @@ contains
       logical, intent(in), optional :: opt_reduce_prec
       TYPE(DECOMP_INFO), target, intent(IN), optional :: opt_decomp
 
+      ! Local variable(s)
       logical :: reduce
       TYPE(DECOMP_INFO), pointer :: decomp
+
+#ifdef PROFILER
+      if (decomp_profiler_io) call decomp_profiler_start("io_write_one")
+#endif
 
       ! Use the provided decomp_info or the default one
       if (present(opt_decomp)) then
@@ -399,6 +425,10 @@ contains
 
       nullify (decomp)
 
+#ifdef PROFILER
+      if (decomp_profiler_io) call decomp_profiler_end("io_write_one")
+#endif
+
    end subroutine write_one_dreal
    !
    subroutine write_one_dcplx(ipencil, var, varname, opt_mode, opt_family, &
@@ -406,6 +436,7 @@ contains
 
       implicit none
 
+      ! Arguments
       integer, intent(IN) :: ipencil
       complex(real64), contiguous, dimension(:, :, :), intent(IN) :: var
[…]
