### 2023-11-06 c18e488
io_read_one : move profiling to the high-level subroutines
