### 2023-11-06 f5a0ba3
io_write_plane : move profiling to the high level subroutines
