+++ src/io.f90
@@ -745,6 +745,7 @@ contains
 
       implicit none
 
+      ! Arguments
       integer, intent(IN) :: ipencil
       real(real32), contiguous, dimension(:, :, :), intent(IN) :: var
       character(len=*), intent(in) :: varname
@@ -756,9 +757,14 @@ contains
       logical, intent(in), optional :: opt_reduce_prec
       TYPE(DECOMP_INFO), target, intent(IN), optional :: opt_decomp
 
+      ! Local variables
       real(real32), allocatable, dimension(:, :, :) :: var2d
       TYPE(DECOMP_INFO), pointer :: decomp
       integer :: nplanes, iplane
+
+#ifdef PROFILER
+      if (decomp_profiler_io) call decomp_profiler_start("io_write_plane")
+#endif
 
       if (present(opt_nplanes)) then
          nplanes = opt_nplanes
@@ -815,6 +821,10 @@ contains
       associate (p => opt_reduce_prec)
       end associate
 
+#ifdef PROFILER
+      if (decomp_profiler_io) call decomp_profiler_end("io_write_plane")
+#endif
+
    end subroutine write_plane_freal
    !
    subroutine write_plane_fcplx(ipencil, var, varname, &
@@ -823,6 +833,7 @@ contains
 
       implicit none
 
+      ! Arguments
       integer, intent(IN) :: ipencil
       complex(real32), contiguous, dimension(:, :, :), intent(IN) :: var
       character(len=*), intent(in) :: varname
@@ -834,9 +845,14 @@ contains
       logical, intent(in), optional :: opt_reduce_prec
       TYPE(DECOMP_INFO), target, intent(IN), optional :: opt_decomp
 
+      ! Local variables
       complex(real32), allocatable, dimension(:, :, :) :: var2d
       TYPE(DECOMP_INFO), pointer :: decomp
       integer :: nplanes, iplane
+
+#ifdef PROFILER
+      if (decomp_profiler_io) call decomp_profiler_start("io_write_plane")
+#endif
 
       if (present(opt_nplanes)) then
          nplanes = opt_nplanes
@@ -893,6 +909,10 @@ contains
       associate (p => opt_reduce_prec)
       end associate
 
+#ifdef PROFILER
+      if (decomp_profiler_io) call decomp_profiler_end("io_write_plane")
+#endif
+
    end subroutine write_plane_fcplx
    !
    subroutine write_plane_dreal(ipencil, var, varname, &
@@ -901,6 +921,7 @@ contains
 
       implicit none
 
+      ! Arguments
       integer, intent(IN) :: ipencil
       real(real64), contiguous, dimension(:, :, :), intent(IN) :: var
       character(len=*), intent(in) :: varname
@@ -912,10 +933,15 @@ contains
       logical, intent(in), optional :: opt_reduce_prec
       TYPE(DECOMP_INFO), target, intent(IN), optional :: opt_decomp
 
+      ! Local variables
       logical :: reduce
       real(real64), allocatable, dimension(:, :, :) :: var2d
       TYPE(DECOMP_INFO), pointer :: decomp
       integer :: nplanes, iplane
+
+#ifdef PROFILER
+      if (decomp_profiler_io) call decomp_profiler_start("io_write_plane")
+#endif
 
       if (present(opt_nplanes)) then
          nplanes = opt_nplanes
@@ -994,6 +1020,10 @@ contains
 
       nullify (decomp)
 
+#ifdef PROFILER
+      if (decomp_profiler_io) call decomp_profiler_end("io_write_plane")
+#endif
+
    end subroutine write_plane_dreal
    !
    subroutine write_plane_dcplx(ipencil, var, varname, &
[…]
