### 2023-11-06 6f4c072
Update HOWTO.md for text corrections
