### 2023-11-06 9e98cbc
Remove some old IO code no longer used
