+++ src/io.f90
@@ -1783,265 +1783,6 @@ contains
 
    end subroutine adios2_read_var
 
-   subroutine write_one_complex(ipencil, var, filename, opt_decomp)
-
-      implicit none
-
-      integer, intent(IN) :: ipencil
-      complex(mytype), contiguous, dimension(:, :, :), intent(IN) :: var
-      character(len=*), intent(IN) :: filename
-      TYPE(DECOMP_INFO), intent(IN), optional :: opt_decomp
-
-      TYPE(DECOMP_INFO) :: decomp
-      integer(kind=MPI_OFFSET_KIND) :: filesize, disp
-      integer, dimension(3) :: sizes, subsizes, starts
-      integer :: ierror, newtype, fh, data_type
-
-#ifdef PROFILER
-      if (decomp_profiler_io) call decomp_profiler_start("io_write_one_cplx")
-#endif
-
-      data_type = complex_type
-
-#include "io_write_one.inc"
-
-#ifdef PROFILER
-      if (decomp_profiler_io) call decomp_profiler_end("io_write_one_cplx")
-#endif
-
-      return
-   end subroutine write_one_complex
-
-!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
-   ! Using MPI-IO library to read from a file a single 3D array
-!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
-   subroutine read_one_real(ipencil, var, dirname, varname, io_name, opt_decomp, reduce_prec)
-
-      implicit none
-
-      integer, intent(IN) :: ipencil
-      real(mytype), contiguous, dimension(:, :, :), intent(INOUT) :: var
-      character(len=*), intent(IN) :: varname, dirname, io_name
-      TYPE(DECOMP_INFO), intent(IN), optional :: opt_decomp
-      logical, intent(in), optional :: reduce_prec
-
-      logical :: read_reduce_prec
-
-      integer :: idx
-#ifndef ADIOS2
-      TYPE(DECOMP_INFO) :: decomp
-      integer, dimension(3) :: sizes, subsizes, starts
-      real(mytype_single), allocatable, dimension(:, :, :) :: varsingle
-      integer :: data_type
-      logical :: dir_exists
-      integer :: disp_bytes
-      integer :: ierror, newtype
-      character(len=:), allocatable :: full_io_name
-      logical :: opened_new
-#endif
-
-#ifdef PROFILER
-      if (decomp_profiler_io) call decomp_profiler_start("io_read_one_real")
-#endif
-
-      read_reduce_prec = .true.
-
-      idx = get_io_idx(io_name, dirname)
-#ifndef ADIOS2
-      opened_new = .false.
-      if (idx < 1) then
-         ! Check file exists
-         full_io_name = trim(dirname)//"/"//trim(varname)
-         if (nrank == 0) then
-            inquire (file=full_io_name, exist=dir_exists)
-            if (.not. dir_exists) then
-               print *, "ERROR: cannot read from", full_io_name, " directory doesn't exist!"
-               stop
-            end if
-         end if
-
-         call decomp_2d_open_io(io_name, full_io_name, decomp_2d_read_mode)
-         idx = get_io_idx(io_name, full_io_name)
-         opened_new = .true.
-      else
-         full_io_name = "" ! Ensure string is not unset
-      end if
-
-      if (present(reduce_prec)) then
-         if (.not. reduce_prec) then
-            read_reduce_prec = .false.
[…]
