+++ examples/grad3d/grad3d.f90
@@ -1,0 +1,379 @@
+!! SPDX-License-Identifier: BSD-3-Clause
+!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
+! This example calculates the gradient of a periodic field using global
+! transposition
+!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
+program grad3d
+
+   use mpi
+
+   use decomp_2d
+   use decomp_2d_constants
+   use decomp_2d_mpi
+   use MPI
+#if defined(_GPU)
+   use cudafor
+   use openacc
+#endif
+
+   implicit none
+
+   integer, parameter :: nx_base = 64, ny_base = 64, nz_base = 64
+   integer :: nx, ny, nz
+   integer :: p_row = 0, p_col = 0
+   integer :: resize_domain
+   integer :: nranks_tot
+   integer :: nargin, arg, FNLength, status, DecInd
+   character(len=80) :: InputFN
+
+   real(mytype), parameter :: lx = 1.0_mytype
+   real(mytype), parameter :: ly = 1.0_mytype
+   real(mytype), parameter :: lz = 1.0_mytype
+
+   real(mytype) :: dx, dy, dz
+
+   real(mytype), allocatable, dimension(:, :, :) :: phi1, phi2, phi3
+   real(mytype), allocatable, dimension(:, :, :) :: dphiX, dphiY, dphiz
+   real(mytype), allocatable, dimension(:, :, :) :: wk2, wk3
+
+   integer :: ierror
+
+   integer :: xlast, ylast, zlast
+
+
+   logical :: passing, all_pass
+
+   call MPI_INIT(ierror)
+   ! To resize the domain we need to know global number of ranks
+   ! This operation is also done as part of decomp_2d_init
+   call MPI_COMM_SIZE(MPI_COMM_WORLD, nranks_tot, ierror)
+   resize_domain = int(nranks_tot / 4) + 1
+   nx = nx_base * resize_domain
+   ny = ny_base * resize_domain
+   nz = nz_base * resize_domain
+   ! Now we can check if user put some inputs
+   ! Handle input file like a boss -- GD
+   nargin = command_argument_count()
+   if ((nargin == 0) .or. (nargin == 2) .or. (nargin == 5)) then
+      do arg = 1, nargin
+         call get_command_argument(arg, InputFN, FNLength, status)
+         read (InputFN, *, iostat=status) DecInd
+         if (arg == 1) then
+            p_row = DecInd
+         elseif (arg == 2) then
+            p_col = DecInd
+         elseif (arg == 3) then
+            nx = DecInd
+         elseif (arg == 4) then
+            ny = DecInd
+         elseif (arg == 5) then
+            nz = DecInd
+         end if
+      end do
+   else
+      ! nrank not yet computed we need to avoid write
+      ! for every rank
+      call MPI_COMM_RANK(MPI_COMM_WORLD, nrank, ierror)
+      if (nrank == 0) then
+         print *, "This Test takes no inputs or 2 inputs as"
+         print *, "  1) p_row (default=0)"
+         print *, "  2) p_col (default=0)"
+         print *, "or 5 inputs as"
+         print *, "  1) p_row (default=0)"
+         print *, "  2) p_col (default=0)"
+         print *, "  3) nx "
+         print *, "  4) ny "
+         print *, "  5) nz "
+         print *, "Number of inputs is not correct and the defult settings"
+         print *, "will be used"
[…]
