+++ examples/grad3d/CMakeLists.txt
@@ -1,0 +1,16 @@
+file(GLOB files_test grad3d.f90)
+include_directories(${CMAKE_SOURCE_DIR}/src)
+
+add_executable(grad3d ${files_test})
+target_link_libraries(grad3d PRIVATE decomp2d)
+
+# Run the test(s)
+set(run_dir "${test_dir}/grad3d")
+message(STATUS "Example dir ${run_dir}")
+file(MAKE_DIRECTORY ${run_dir})
+if (BUILD_TARGET MATCHES "gpu")
+  file(COPY bind.sh DESTINATION ${run_dir})
+  add_test(NAME grad3d COMMAND ${MPIEXEC_EXECUTABLE} ${MPIEXEC_NUMPROC_FLAG} ${MPIEXEC_MAX_NUMPROCS} ./bind.sh $<TARGET_FILE:grad3d> ${TEST_ARGUMENTS} WORKING_DIRECTORY ${run_dir})
+else ()  
+  add_test(NAME grad3d COMMAND ${MPIEXEC_EXECUTABLE} ${MPIEXEC_NUMPROC_FLAG} ${MPIEXEC_MAX_NUMPROCS} $<TARGET_FILE:grad3d> ${TEST_ARGUMENTS} WORKING_DIRECTORY ${run_dir})
+endif ()
