+++ examples/CMakeLists.txt
@@ -12,3 +12,4 @@ add_subdirectory(fft_physical_x)
 add_subdirectory(fft_physical_z)
 add_subdirectory(halo_test)
 add_subdirectory(io_test)
+add_subdirectory(grad3d)
