### 2023-11-07 14f8a88
Add visualisation to grad3d example
