+++ examples/grad3d/grad3d.f90
@@ -24,6 +24,8 @@ program grad3d
    integer :: resize_domain
    integer :: nranks_tot
    integer :: nargin, arg, FNLength, status, DecInd
+   integer :: ierror
+   logical :: passing, all_pass
    character(len=80) :: InputFN
 
    real(mytype), parameter :: lx = 1.0_mytype
@@ -31,17 +33,12 @@ program grad3d
    real(mytype), parameter :: lz = 1.0_mytype
 
    real(mytype) :: dx, dy, dz
+   real(mytype) :: error_ref 
 
    real(mytype), allocatable, dimension(:, :, :) :: phi1, phi2, phi3
    real(mytype), allocatable, dimension(:, :, :) :: dphiX, dphiY, dphiz
    real(mytype), allocatable, dimension(:, :, :) :: wk2, wk3
 
-   integer :: ierror
-
-   integer :: xlast, ylast, zlast
-
-
-   logical :: passing, all_pass
 
    call MPI_INIT(ierror)
    ! To resize the domain we need to know global number of ranks
@@ -93,14 +90,21 @@ program grad3d
    dx = lx/real(nx,mytype)
    dy = ly/real(ny,mytype)
    dz = lz/real(nz,mytype)
+   
+   if (nrank == 0) then
+      write (*, *) '-----------------------------------------------'
+      write (*, *) "Mesh Resolution ", nx , ny, nz 
+      write (*, *) '-----------------------------------------------'
+   end if
 
    call allocate_var()
    !$acc data create(phi2,phi3,wk2,wk3) copy(phi1,dphiX, dphiY, dphiZ)
    call init_phi()
    call compute_grad()
    call test_derX(dphiX)
-   call test_derY()
-   call test_derZ()
+   call test_derY(dphiY)
+   call test_derZ(dphiZ)
+   call write_data()
    !$acc end data
 
    if (nrank == 0) then
@@ -207,7 +211,7 @@ contains
 
       ! Compute Z derivative
       call transpose_y_to_z(phi2,phi3)
-      call derx(wk3,phi3,dz,zsize(1),zsize(2),zsize(3))
+      call derz(wk3,phi3,dz,zsize(1),zsize(2),zsize(3))
       call transpose_z_to_y(wk3,wk2)
       call transpose_y_to_x(wk2,dphiZ)
 
@@ -323,6 +327,8 @@ contains
       real(mytype) :: dphi, dphi_num
       real(mytype) :: error = 0._mytype
       real(mytype) :: err_all = 0._mytype
+      real(mytype) :: dphi2 = 0._mytype
+      real(mytype) :: sum_dphi2 = 0._mytype
       integer :: xe1, xe2, xe3
       integer :: xs1, xs2, xs3
 
@@ -341,7 +347,8 @@ contains
              z = (k + xs3 - 2) * dz
              y = (j + xs2 - 2) * dy
              x = (i + xs1 - 2) * dx
-             dphi = -2._mytype * twopi * sin (twopi * (x / lx)) * cos(twopi * (y / ly)) * sin(twopi * (z / lz))
+             dphi = -2._mytype * twopi * sin(twopi * (x / lx)) * cos(twopi * (y / ly)) * sin(twopi * (z / lz))
+             dphi2 = dphi2 + dphi * dphi
              dphi_num = df(i,j,k)
              error = error + (dphi-dphi_num)*(dphi-dphi_num)
           enddo
@@ -349,8 +356,10 @@ contains
       enddo
       !$acc end loop
       call MPI_ALLREDUCE(error, err_all, 1, real_type, MPI_SUM, MPI_COMM_WORLD, ierror)
[…]
