### 2023-11-07 ee4c043
Remove unused variable
