### 2023-11-07 cbd856c
Typos correction for grad3d example
