### 2023-11-07 6ee1748
Correct analytical solution adding domain lenght
