### 2023-11-07 f2fe504
Add io initialisation to grad3d
