+++ examples/grad3d/CMakeLists.txt
@@ -8,6 +8,7 @@ target_link_libraries(grad3d PRIVATE decomp2d)
 set(run_dir "${test_dir}/grad3d")
 message(STATUS "Example dir ${run_dir}")
 file(MAKE_DIRECTORY ${run_dir})
+file (COPY "${CMAKE_SOURCE_DIR}/examples/io_test/adios2_config.xml" DESTINATION ${run_dir})
 if (BUILD_TARGET MATCHES "gpu")
   file(COPY bind.sh DESTINATION ${run_dir})
   add_test(NAME grad3d COMMAND ${MPIEXEC_EXECUTABLE} ${MPIEXEC_NUMPROC_FLAG} ${MPIEXEC_MAX_NUMPROCS} ./bind.sh $<TARGET_FILE:grad3d> ${TEST_ARGUMENTS} WORKING_DIRECTORY ${run_dir})
