### 2023-11-07 02f2caa
Add io initialisation to grad3d
