### 2023-11-07 e939cb9
Update HOWTO.md
Add documentation on I/O initialisation
