### 2023-11-07 eec7fe1
Include GPU section in the JOSS paper
