### 2023-11-07 3135c8a
Update GPU section in the JOSS paper
