+++ submissionjoss/paper.md
@@ -99,17 +99,17 @@ The new library was designed to be more appealing to the scientific community,
 and can now be easily implemented as an independent library for use by other software.
 
 # GPU porting
-A first attempt to port 2DECOMP&FFT to GPU
-has been performed within the solver AFiD-GPU [@ZHU2018199], which was mainly based 
-on CUDA-Fortran for some kernels and CUDA-aware-MPI for communication.
-A second library, named cuDECOMP, which is directly inspired by 2DECOMP&FFT, takes 
-full adavantages of CUDA and also uses NVIDIA most recent libraries for communication 
-such as NVIDIA Collective Communication Library (NCCL), is presented in [@Romero_2022_cuDecomp]. 
-This work mainly uses a mix of CUDA-fortran and openACC for the CPU porting together with 
-CUDA-aware-MPI and NCCL for the communications. In additions to previous work also the FFT module 
-is ported to GPs using CUDA-FFT. 
-The long term goal is also to introduce openMP for GPU porting to support both AMD and Intel 
-GPU hardware. 
+
+An initial port of 2DECOMP&FFT to GPU has been performed within the solver AFiD-GPU [@ZHU2018199],
+which was mainly based on CUDA-Fortran for some kernels and CUDA-aware-MPI for communications.
+A second library, named cuDECOMP, which is directly inspired by 2DECOMP&FFT, 
+takes full adavantages of CUDA and also uses NVIDIA most recent libraries for communications 
+such as NVIDIA Collective Communication Library (NCCL), is presented in [@Romero_2022_cuDecomp].
+Indeed, cuDECOMP only targets NVIDIA GPUs.
+The updated 2DECOMP&FFT mainly uses a mix of CUDA-fortran and openACC for the GPU porting 
+together with CUDA-aware-MPI and NCCL for the communications. In additions to previous work,
+the FFT module is ported to GPUs using CUDA-FFT. 
+The next step is also to use openMP for GPU porting to support both AMD and Intel GPU hardware.
 
 # How to use 2DECOMP&FFT
 The 2D Pencil Decomposition API is defined with three Fortran module which should be used by applications as:
