### 2023-11-07 57c6f18
Add output2D definition for xdmf output in grad3d example
