### 2023-11-07 1930887
Typos correction for JOSS submission
