+++ submissionjoss/paper.md
@@ -63,18 +63,18 @@ with excellent strong scaling performance up to hundreds of thousands of CPU cor
 2DECOMP&FFT mainly relies on MPI, and it offers a user-friendly interface that hides the complexity 
 of the communication. 
 The version 2.0.0 of the library offers also a 1D slab decomposition, which is implemented as a special case of the 
-2D decomposition. Two alteratives are possible:
+2D decomposition. Two alternatives are possible:
 
 - Initial slabs orientation in the `XY` plane; 
 - Initial slabs orientation in the `XZ` plane.
 
-In many configurations the slabs decompostion gives some gain in performances with respect to the 
-2D  pencil decomposion. This is a consequence of having data already in memory when transposing between 
-the two dimensions of the slab. Therefore, it is possible to perfor a simple memory copy between 
+In many configurations the slabs decomposition gives some gain in performances with respect to the 
+2D-pencil decomposition. This is a consequence of having data already in memory when transposing between 
+the two dimensions of the slab. Therefore, it is possible to perform a simple memory copy between 
 input and output arrays instead of the full MPI communication.
 
 The library also offers also a very efficient and flexible 3D tool to perform 
-Fast Fourier Transform (FFT) for distrubuted memory systems. However, 2DECOMP&FFT is mainly designed to perform 
+Fast Fourier Transform (FFT) for distributed memory systems. However, 2DECOMP&FFT is mainly designed to perform 
 management and communication and the actual computation of the 1D FFT is delegated to 3rd-parly libraries. 
 The supported FFT backends are: FFTW [@FFTW05], the Intel Math Kernel Library (MKL) and the CUDA FFT (cuFFT) 
 which is used for FFT on NVIDIA GPUs. A Generic FFT backend, based on the 
@@ -85,7 +85,7 @@ While 2DECOMP&FFT library has been designed with high order compact schemes in m
 that some derivatives can be evaluated using explicit formulation using local stencils. For this reason a 
 halo support API is also provided to support explicit message passing between neighbouring pencils. 
 
-Finally the library provide an infrastucture to perfrom data I/O using MPI I/O or ADIOS2 [@godoy2020adios]. 
+Finally, the library provides an infrastructure to perform data I/O using MPI I/O or ADIOS2 [@godoy2020adios]. 
 The API provide several features such as: write of a single or multiple 3D arrays into a file, write of 2D slices 
 from a 2D array, write of the results using a lower precision (i.e. double precision into single) 
 and with lower resolution. 
@@ -103,10 +103,10 @@ and can now be easily implemented as an independent library for use by other sof
 An initial port of 2DECOMP&FFT to GPU has been performed within the solver AFiD-GPU [@ZHU2018199],
 which was mainly based on CUDA-Fortran for some kernels and CUDA-aware-MPI for communications.
 A second library, named cuDECOMP, which is directly inspired by 2DECOMP&FFT, 
-takes full adavantages of CUDA and also uses NVIDIA most recent libraries for communications 
+takes full adavantages of CUDA and uses NVIDIA most recent libraries for communications 
 such as NVIDIA Collective Communication Library (NCCL), is presented in [@Romero_2022_cuDecomp].
 Indeed, cuDECOMP only targets NVIDIA GPUs.
-The updated 2DECOMP&FFT mainly uses a mix of CUDA-fortran and openACC for the GPU porting 
+The updated 2DECOMP&FFT mainly uses a mix of CUDA-Fortran and openACC for the GPU porting 
 together with CUDA-aware-MPI and NCCL for the communications. In additions to previous work,
 the FFT module is ported to GPUs using CUDA-FFT. 
[…]
