### 2023-11-08 dfbb9d3
Propose changes to paper.md
