### 2023-11-08 4197ce1
Further corrections to JOSS paper
