### 2023-11-08 e6a7508
Change intent for gradient in test routines  under the grad3d example
