### 2023-11-08 085d42a
Add the link to the example README under JOSS paper
