### 2023-11-08 a4a4168
Add Grad3d test to list of example and perform fprettify
