+++ examples/grad3d/grad3d.f90
@@ -33,12 +33,11 @@ program grad3d
    real(mytype), parameter :: lz = 1.0_mytype
 
    real(mytype) :: dx, dy, dz
-   real(mytype) :: error_ref 
+   real(mytype) :: error_ref
 
    real(mytype), allocatable, dimension(:, :, :) :: phi1, phi2, phi3
    real(mytype), allocatable, dimension(:, :, :) :: dphiX, dphiY, dphiz
    real(mytype), allocatable, dimension(:, :, :) :: wk2, wk3
-
 
    call MPI_INIT(ierror)
    ! To resize the domain we need to know global number of ranks
@@ -87,13 +86,13 @@ program grad3d
    end if
    call decomp_2d_init(nx, ny, nz, p_row, p_col)
 
-   dx = lx/real(nx,mytype)
-   dy = ly/real(ny,mytype)
-   dz = lz/real(nz,mytype)
-   
+   dx = lx / real(nx, mytype)
+   dy = ly / real(ny, mytype)
+   dz = lz / real(nz, mytype)
+
    if (nrank == 0) then
       write (*, *) '-----------------------------------------------'
-      write (*, *) "Mesh Resolution ", nx , ny, nz 
+      write (*, *) "Mesh Resolution ", nx, ny, nz
       write (*, *) '-----------------------------------------------'
    end if
 
@@ -159,7 +158,7 @@ contains
       implicit none
 
       integer :: i, j, k
-      real(mytype), parameter :: twopi=2._mytype*acos(-1._mytype)
+      real(mytype), parameter :: twopi = 2._mytype * acos(-1._mytype)
       real(mytype) :: x, y, z
       integer :: xe1, xe2, xe3
       integer :: xs1, xs2, xs3
@@ -173,12 +172,12 @@ contains
 
       ! Scalar field
       !$acc kernels default(present)
-      do concurrent (k=1:xe3, j=1:xe2, i=1:xe1)
+      do concurrent(k=1:xe3, j=1:xe2, i=1:xe1)
          z = (k + xs3 - 2) * dz
          y = (j + xs2 - 2) * dy
          x = (i + xs1 - 2) * dx
          phi1(i, j, k) = -2._mytype * cos(twopi * (x / lx)) * cos(twopi * (y / ly)) * sin(twopi * (z / lz))
-      enddo
+      end do
       !$acc end kernels
 
    end subroutine init_phi
@@ -202,116 +201,116 @@ contains
       implicit none
 
       ! Compute X derivative
-      call derx(dphiX,phi1,dx,xsize(1),xsize(2),xsize(3))
+      call derx(dphiX, phi1, dx, xsize(1), xsize(2), xsize(3))
 
       ! Compute Y derivative
-      call transpose_x_to_y(phi1,phi2)
-      call dery(wk2,phi2,dy,ysize(1),ysize(2),ysize(3))
-      call transpose_y_to_x(wk2,dphiY)
+      call transpose_x_to_y(phi1, phi2)
+      call dery(wk2, phi2, dy, ysize(1), ysize(2), ysize(3))
+      call transpose_y_to_x(wk2, dphiY)
 
       ! Compute Z derivative
-      call transpose_y_to_z(phi2,phi3)
-      call derz(wk3,phi3,dz,zsize(1),zsize(2),zsize(3))
-      call transpose_z_to_y(wk3,wk2)
-      call transpose_y_to_x(wk2,dphiZ)
+      call transpose_y_to_z(phi2, phi3)
+      call derz(wk3, phi3, dz, zsize(1), zsize(2), zsize(3))
+      call transpose_z_to_y(wk3, wk2)
+      call transpose_y_to_x(wk2, dphiZ)
 
    end subroutine compute_grad
    !=====================================================================
[…]
