+++ examples/README.md
@@ -6,5 +6,6 @@
 - [Test FFT from Z pencil](fft_physical_z) - various tests for the FFT starting from the ``Z`` direction 
 - [Test HALO exchange](halo_test)          - to test the halo-cell exchange capability
 - [Test I/O features](io_test)             - various tests for the I/O module
+- [Example gradient of a scalar](grad3d)   - example of how to compute the gradient of a scalar field
 
 Refer to the README files for each example for details.
