### 2023-11-08 #5813 @rfj82982
@diehlpk @CeciliaCoelho  @p-costa Thanks very much for your reviews. We have a new version of the main branch of 2DECOMP&FFT and related new version of the article. Here a list of the modifications done to meet the requests from reviewers:

-  At the top of the [README](https://github.com/2decomp-fft/2decomp-fft/blob/main/README.md) file some new text has been added pointing to a new file named [HOWTO.md](https://github.com/2decomp-fft/2decomp-fft/blob/main/HOWTO.md). The file gives a brief overview on how to include 2DECOMP&FFT routines into a code. The files points also to the github.io site with the more in-depth description of all API. A slim down version of the file has also been included into the paper under a dedicated [section](https://github.com/2decomp-fft/2decomp-fft/blob/277eb34cfde33be4fcfcda038770f074fc919cb9/submissionjoss/paper.md?plain=1#L114)
- At the top of the [README](https://github.com/2decomp-fft/2decomp-fft/blob/main/README.md) a link to the example directory is provided. The description of all examples has been updated to describe the feature of every single example. The example mentioned about the calculation of the divergence is under [Halo](https://github.com/2decomp-fft/2decomp-fft/blob/main/examples/halo_test/halo_test.f90) since it is used to test the halo functionality of the library
- A new example that compute the [gradient](https://github.com/2decomp-fft/2decomp-fft/blob/main/examples/grad3d/grad3d.f90) of a scalar field has also been included. This is to show what a full solver will look like using both transposes routines and I/O. The gradients are very simple second order central scheme in all 3 directions on a periodic case. Link to the example directory of the repository has also been included.
- The how to [contribute](https://github.com/2decomp-fft/2decomp-fft/blob/main/Contribute.md) file has now been linked at the bottom of the README file.
- A new section related to the GPU implementation has been included directly in the [paper](https://github.com/2decomp-fft/2decomp-fft/blob/277eb34cfde33be4fcfcda038770f074fc919cb9/submissionjoss/paper.md?plain=1#L101) with reference to previous works.
- The paper has been checked few times for grammar and typos.

Please let us know if this address all the points raised in both reviews.

Thanks

Stefano, Cedric, Paul, Filippo and Sylvain
