### 2023-11-08 #5813 @p-costa
Hi Stefano,

Thanks for considering my suggestions. I looked into the revised manuscript and the changes, and I have no further comments. I think 2decomp has had a major impact on many scientific studies, and I am happy to see the project continuing to thrive.
