+++ src/transpose_z_to_y.f90
@@ -4,484 +4,484 @@
 submodule(decomp_2d) d2d_transpose_z_to_y
 
 #ifdef NVHPC
-  ! FIXME : Nvidia compiler does not compile the submodule without this
-  use decomp_2d_constants, only: mytype
+   ! FIXME : Nvidia compiler does not compile the submodule without this
+   use decomp_2d_constants, only: mytype
 #endif
 
-  implicit none
+   implicit none
 
 contains
 
-  module subroutine transpose_z_to_y_real_short(src, dst)
+   module subroutine transpose_z_to_y_real_short(src, dst)
 
-    implicit none
+      implicit none
 
-    real(mytype), dimension(:, :, :), intent(IN) :: src
-    real(mytype), dimension(:, :, :), intent(OUT) :: dst
+      real(mytype), dimension(:, :, :), intent(IN) :: src
+      real(mytype), dimension(:, :, :), intent(OUT) :: dst
 
-    call transpose_z_to_y(src, dst, decomp_main)
+      call transpose_z_to_y(src, dst, decomp_main)
 
-  end subroutine transpose_z_to_y_real_short
+   end subroutine transpose_z_to_y_real_short
 
-  module subroutine transpose_z_to_y_real_long(src, dst, decomp)
+   module subroutine transpose_z_to_y_real_long(src, dst, decomp)
 
-    implicit none
+      implicit none
 
-    real(mytype), dimension(:, :, :), intent(IN) :: src
-    real(mytype), dimension(:, :, :), intent(OUT) :: dst
-    TYPE(DECOMP_INFO), intent(IN) :: decomp
+      real(mytype), dimension(:, :, :), intent(IN) :: src
+      real(mytype), dimension(:, :, :), intent(OUT) :: dst
+      TYPE(DECOMP_INFO), intent(IN) :: decomp
 
 #if defined(_GPU)
-    integer :: istat, nsize
+      integer :: istat, nsize
 #endif
 
 #ifdef PROFILER
-    if (decomp_profiler_transpose) call decomp_profiler_start("transp_z_y_r")
+      if (decomp_profiler_transpose) call decomp_profiler_start("transp_z_y_r")
 #endif
 
 #if defined(_GPU)
-    integer :: istat, nsize
+      integer :: istat, nsize
 #endif
 
-    if (dims(2) == 1) then
+      if (dims(2) == 1) then
 #if defined(_GPU)
-       nsize = product(decomp%zsz)
-       !$acc host_data use_device(src,dst)
-       istat = cudaMemcpy(dst, src, nsize, cudaMemcpyDeviceToDevice)
-       !$acc end host_data
-       if (istat /= 0) call decomp_2d_abort(__FILE__, __LINE__, istat, "cudaMemcpy")
+         nsize = product(decomp%zsz)
+         !$acc host_data use_device(src,dst)
+         istat = cudaMemcpy(dst, src, nsize, cudaMemcpyDeviceToDevice)
+         !$acc end host_data
+         if (istat /= 0) call decomp_2d_abort(__FILE__, __LINE__, istat, "cudaMemcpy")
 #else
-       dst = src
+         dst = src
 #endif
-    else
+      else
 #if defined(_GPU)
-       call transpose_z_to_y_real(src, dst, decomp, work1_r_d, work2_r_d)
+         call transpose_z_to_y_real(src, dst, decomp, work1_r_d, work2_r_d)
 #else
-       call transpose_z_to_y_real(src, dst, decomp, work1_r, work2_r)
+         call transpose_z_to_y_real(src, dst, decomp, work1_r, work2_r)
 #endif
-    end if
+      end if
 
[…]
