+++ src/transpose_y_to_x.f90
@@ -4,435 +4,435 @@
 submodule(decomp_2d) d2d_transpose_y_to_x
 
 #ifdef NVHPC
-  ! FIXME : Nvidia compiler does not compile the submodule without this
-  use decomp_2d_constants, only: mytype
+   ! FIXME : Nvidia compiler does not compile the submodule without this
+   use decomp_2d_constants, only: mytype
 #endif
 
-  implicit none
+   implicit none
 
 contains
 
-  module subroutine transpose_y_to_x_real_short(src, dst)
+   module subroutine transpose_y_to_x_real_short(src, dst)
 
-    implicit none
+      implicit none
 
-    real(mytype), dimension(:, :, :), intent(IN) :: src
-    real(mytype), dimension(:, :, :), intent(OUT) :: dst
+      real(mytype), dimension(:, :, :), intent(IN) :: src
+      real(mytype), dimension(:, :, :), intent(OUT) :: dst
 
-    call transpose_y_to_x(src, dst, decomp_main)
+      call transpose_y_to_x(src, dst, decomp_main)
 
-  end subroutine transpose_y_to_x_real_short
+   end subroutine transpose_y_to_x_real_short
 
-  module subroutine transpose_y_to_x_real_long(src, dst, decomp)
+   module subroutine transpose_y_to_x_real_long(src, dst, decomp)
 
-    implicit none
+      implicit none
 
-    real(mytype), dimension(:, :, :), intent(IN) :: src
-    real(mytype), dimension(:, :, :), intent(OUT) :: dst
-    TYPE(DECOMP_INFO), intent(IN) :: decomp
+      real(mytype), dimension(:, :, :), intent(IN) :: src
+      real(mytype), dimension(:, :, :), intent(OUT) :: dst
+      TYPE(DECOMP_INFO), intent(IN) :: decomp
 
 #if defined(_GPU)
-    integer :: istat, nsize
+      integer :: istat, nsize
 #endif
 
 #ifdef PROFILER
-    if (decomp_profiler_transpose) call decomp_profiler_start("transp_y_x_r")
+      if (decomp_profiler_transpose) call decomp_profiler_start("transp_y_x_r")
 #endif
 
-    if (dims(1) == 1) then
+      if (dims(1) == 1) then
 #if defined(_GPU)
-       nsize = product(decomp%ysz)
-       !$acc host_data use_device(src,dst)
-       istat = cudaMemcpy(dst, src, nsize, cudaMemcpyDeviceToDevice)
-       !$acc end host_data
-       if (istat /= 0) call decomp_2d_abort(__FILE__, __LINE__, istat, "cudaMemcpy")
+         nsize = product(decomp%ysz)
+         !$acc host_data use_device(src,dst)
+         istat = cudaMemcpy(dst, src, nsize, cudaMemcpyDeviceToDevice)
+         !$acc end host_data
+         if (istat /= 0) call decomp_2d_abort(__FILE__, __LINE__, istat, "cudaMemcpy")
 #else
-       dst = src
+         dst = src
 #endif
-    else
+      else
 #if defined(_GPU)
-       call transpose_y_to_x_real(src, dst, decomp, work1_r_d, work2_r_d)
+         call transpose_y_to_x_real(src, dst, decomp, work1_r_d, work2_r_d)
 #else
-       call transpose_y_to_x_real(src, dst, decomp, work1_r, work2_r)
+         call transpose_y_to_x_real(src, dst, decomp, work1_r, work2_r)
 #endif
-    end if
+      end if
 
 #ifdef PROFILER
[…]

[…]
