### 2023-11-09 41a9d74
WIP: tidy transpose_z_to_y
