+++ src/transpose_z_to_y.f90
@@ -31,30 +31,25 @@ contains
       real(mytype), dimension(:, :, :), intent(OUT) :: dst
       TYPE(DECOMP_INFO), intent(IN) :: decomp
 
-#if defined(_GPU)
-      integer :: istat, nsize
-#endif
-
+      real(mytype), dimension(:), pointer, contiguous :: wk1
+      real(mytype), dimension(:), pointer, contiguous :: wk2
+      
 #ifdef PROFILER
       if (decomp_profiler_transpose) call decomp_profiler_start("transp_z_y_r")
 #endif
 
+#if defined(_GPU)
+      wk1 => work1_r_d
+      wk2 => work2_r_d
+#else
+      wk1 => work1_r
+      wk2 => work2_r
+#endif
+
       if (dims(2) == 1) then
-#if defined(_GPU)
-         nsize = product(decomp%zsz)
-         !$acc host_data use_device(src,dst)
-         istat = cudaMemcpy(dst, src, nsize, cudaMemcpyDeviceToDevice)
-         !$acc end host_data
-         if (istat /= 0) call decomp_2d_abort(__FILE__, __LINE__, istat, "cudaMemcpy")
-#else
-         dst = src
-#endif
+         call transpose_z_to_y_real_fast(src, dst, decomp)
       else
-#if defined(_GPU)
-         call transpose_z_to_y_real(src, dst, decomp, work1_r_d, work2_r_d)
-#else
-         call transpose_z_to_y_real(src, dst, decomp, work1_r, work2_r)
-#endif
+         call transpose_z_to_y_real(src, dst, decomp, wk1, wk2)
       end if
 
 #ifdef PROFILER
@@ -63,6 +58,31 @@ contains
 
    end subroutine transpose_z_to_y_real_long
 
+   subroutine transpose_z_to_y_real_fast(src, dst, decomp)
+
+      implicit none
+
+      real(mytype), dimension(:, :, :), intent(IN) :: src
+      real(mytype), dimension(:, :, :), intent(OUT) :: dst
+      TYPE(DECOMP_INFO), intent(IN) :: decomp
+
+      integer :: nsize
+#if defined(_GPU)
+      integer :: istat
+#endif
+
+      nsize = product(decomp%zsz)
+#if defined(_GPU)
+      !$acc host_data use_device(src,dst)
+      istat = cudaMemcpy(dst, src, nsize, cudaMemcpyDeviceToDevice)
+      !$acc end host_data
+      if (istat /= 0) call decomp_2d_abort(__FILE__, __LINE__, istat, "cudaMemcpy")
+#else
+      dst = src
+#endif
+
+   end subroutine transpose_z_to_y_real_fast
+   
    subroutine transpose_z_to_y_real(src, dst, decomp, wk1, wk2)
 
       implicit none
@@ -73,84 +93,14 @@ contains
       real(mytype), dimension(:), intent(out) :: wk1, wk2
 #if defined(_GPU)
       attributes(device) :: wk1, wk2
-
-      integer :: istat
 #endif
 
-      integer :: s1, s2, s3, d1, d2, d3
-      integer :: ierror
-
-      s1 = SIZE(src, 1)
-      s2 = SIZE(src, 2)
-      s3 = SIZE(src, 3)
-      d1 = SIZE(dst, 1)
-      d2 = SIZE(dst, 2)
-      d3 = SIZE(dst, 3)
-
-      ! rearrange source array as send buffer
-#ifdef EVEN
-      if (.not. decomp%even) then
-         call mem_split_zy_real(src, s1, s2, s3, wk1, dims(2), &
-                                decomp%z2dist, decomp)
-      end if
-#else
-      ! note the src array is suitable to be a send buffer
-      ! so no split operation needed
-
-#if defined(_GPU)
[…]
