### 2023-11-09 8ed9249
Correct some ifdef conditions
