### 2023-11-09 82957fb
Fix missing device attribute for _GPU
