### 2023-11-09 49f1087
Fix missing variable declarations for GPU builds
