### 2023-11-09 54f866c
Silence warnings about comments inline with preprocessor directives
