+++ CHANGELOG.md
@@ -1,0 +1,20 @@
+# Changelog
+
+This file tracks the notable changes to the 2decomp&fft project (starting from v2.0).
+
+The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
+and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
+
+When a release is made the "Unreleased" section should be renamed to that version and date
+_e.g._ vX.Y - YYYY-MM-DD and a new "Unreleased" section started above.
+
+## Unreleased
+
+### Added
+### Fixed
+### Changed
+### Removed
+
+## v2.0
+
+- See JOSS paper (under submission)
