### 2023-11-10 b605eb3
Add missing Deprecated section in CHANGELOG
