### 2023-11-10 01b8dc1
Update CHANGELOG.md
Correct release version for JOSS paper
