### 2023-11-10 82d16f6
Delete reference to semver in CHANGELOG
We don't use this
