### 2023-11-10 ae0c90c
Combine NEWS and CHANGELOG
