+++ NEWS.md
@@ -1,13 +1,0 @@
-Dev (not on main branch yet)
-------------------------------
-
-### Transpose
-
-- Fuse CPU and GPU memory buffers to reduce the memory usage. See [PR 271](https://github.com/2decomp-fft/2decomp-fft/pull/271)
-
-### Profiler
-
-- Transpose subroutines are now profiled when `p_row` or `p_col` is equal to one. See [PR #251](https://github.com/2decomp-fft/2decomp-fft/pull/251)
-
-Release 2.0 (main branch on 2023-08-30)
----------------------------------------
