### 2023-11-10 9c5869a
Combine NEWS and CHANGELOG (#295)
* Combine NEWS and CHANGELOG

* Update CHANGELOG.md

---------

Co-authored-by: CFLAG <mathrack@users.noreply.github.com>
