### 2023-11-10 08ed65d
Adapt io_tmp_test to the new IO framework
