+++ examples/io_test/io_tmp_test.f90
@@ -6,11 +6,13 @@
 program io_test
 
    use mpi
+   use decomp_2d
    use decomp_2d_constants
    use decomp_2d_io
+   use decomp_2d_io_family
+   use decomp_2d_io_object
    use decomp_2d_mpi
    use decomp_2d_testing
-   use decomp_2d
 #if defined(_GPU)
    use cudafor
    use openacc
@@ -50,6 +52,8 @@ program io_test
 
    character(len=*), parameter :: io_name = "test-io"
    character(len=*), parameter :: io_restart = "restart-io"
+   type(d2d_io_family), save :: io_family_restart
+   type(d2d_io), save :: io
 
    integer :: i, j, k, m, ierror
    integer :: xst1, xst2, xst3
@@ -77,20 +81,20 @@ program io_test
    call decomp_2d_testing_log()
 
    call decomp_2d_io_init()
-   call decomp_2d_init_io(io_name)
-   call decomp_2d_register_variable(io_name, "u1.dat", 1, 0, output2D, mytype)
-   call decomp_2d_register_variable(io_name, "u2.dat", 2, 0, output2D, mytype)
-   call decomp_2d_register_variable(io_name, "u3.dat", 3, 0, output2D, mytype)
-   call decomp_2d_register_variable(io_name, "v1.dat", 1, 0, output2D, mytype)
-   call decomp_2d_register_variable(io_name, "v2.dat", 2, 0, output2D, mytype)
-   call decomp_2d_register_variable(io_name, "v3.dat", 3, 0, output2D, mytype)
-   call decomp_2d_init_io(io_restart)
-   call decomp_2d_register_variable(io_restart, "u1.dat", 1, 0, output2D, mytype)
-   call decomp_2d_register_variable(io_restart, "u2.dat", 2, 0, output2D, mytype)
-   call decomp_2d_register_variable(io_restart, "u3.dat", 3, 0, output2D, mytype)
-   call decomp_2d_register_variable(io_restart, "v1.dat", 1, 0, output2D, mytype)
-   call decomp_2d_register_variable(io_restart, "v2.dat", 2, 0, output2D, mytype)
-   call decomp_2d_register_variable(io_restart, "v3.dat", 3, 0, output2D, mytype)
+
+   call decomp_2d_io_register_var3d("u1.dat", 1, real_type)
+   call decomp_2d_io_register_var3d("u2.dat", 2, real_type)
+   call decomp_2d_io_register_var3d("u3.dat", 3, real_type)
+   call decomp_2d_io_register_var3d("v1.dat", 1, real_type)
+   call decomp_2d_io_register_var3d("v2.dat", 2, real_type)
+   call decomp_2d_io_register_var3d("v3.dat", 3, real_type)
+   call io_family_restart%init(io_restart)
+   call io_family_restart%register_var3d("u1.dat", 1, real_type)
+   call io_family_restart%register_var3d("u2.dat", 2, real_type)
+   call io_family_restart%register_var3d("u3.dat", 3, real_type)
+   call io_family_restart%register_var3d("v1.dat", 1, real_type)
+   call io_family_restart%register_var3d("v2.dat", 2, real_type)
+   call io_family_restart%register_var3d("v3.dat", 3, real_type)
 
    ! ***** global data *****
    allocate (data1(nx, ny, nz))
[…]
