### 2023-11-11 311b90d
Use the log instead of writing to stdout
