### 2023-11-11 221cb7b
Rename d2d_listing => d2d_log. Add a utility function to check if the rank will write to the log.
