+++ src/log.f90
@@ -1,6 +1,6 @@
 !! SPDX-License-Identifier: BSD-3-Clause
 
-submodule(decomp_2d) d2d_log
+submodule(decomp_2d) d2d_log_submodule
 
    use decomp_2d_constants
    use decomp_2d_mpi
@@ -10,16 +10,38 @@ submodule(decomp_2d) d2d_log
 contains
 
    !
+   ! Return true if this rank will write to the listing
+   !
+   module function d2d_log_is_active()
+
+      implicit none
+
+      ! Output
+      logical :: d2d_log_is_active
+
+      if (decomp_log == D2D_LOG_STDOUT .and. nrank == 0) then
+         d2d_log_is_active = .true.
+      else if (decomp_log == D2D_LOG_TOFILE .and. nrank == 0) then
+         d2d_log_is_active = .true.
+      else if (decomp_log == D2D_LOG_TOFILE_FULL) then
+         d2d_log_is_active = .true.
+      else
+         d2d_log_is_active = .false.
+      end if
+
+   end function d2d_log_is_active
+
+   !
    ! Get the IO unit for the log
    !
-   module function d2d_listing_get_unit()
+   module function d2d_log_get_unit()
 
       use iso_fortran_env, only: output_unit
 
       implicit none
 
       ! Output
-      integer :: d2d_listing_get_unit
+      integer :: d2d_log_get_unit
 
       ! Local variables
       logical :: found
@@ -31,13 +53,13 @@ contains
          inquire (file='decomp_2d_setup_'//trim(fname)//'.log', &
                   exist=found)
          if (found) then
-            open (newunit=d2d_listing_get_unit, &
+            open (newunit=d2d_log_get_unit, &
                   file='decomp_2d_setup_'//trim(fname)//'.log', &
                   status="old", &
                   position="append", &
                   iostat=ierror)
          else
-            open (newunit=d2d_listing_get_unit, &
+            open (newunit=d2d_log_get_unit, &
                   file='decomp_2d_setup_'//trim(fname)//'.log', &
                   status="new", &
                   iostat=ierror)
@@ -46,29 +68,29 @@ contains
          inquire (file="decomp_2d_setup.log", &
                   exist=found)
          if (found) then
-            open (newunit=d2d_listing_get_unit, &
+            open (newunit=d2d_log_get_unit, &
                   file="decomp_2d_setup.log", &
                   status="old", &
                   position="append", &
                   iostat=ierror)
          else
-            open (newunit=d2d_listing_get_unit, &
+            open (newunit=d2d_log_get_unit, &
                   file="decomp_2d_setup.log", &
                   status="new", &
                   iostat=ierror)
          end if
       else
-         d2d_listing_get_unit = output_unit
+         d2d_log_get_unit = output_unit
          ierror = 0
       end if
       if (ierror /= 0) call decomp_2d_abort(__FILE__, __LINE__, ierror, "Could not open log file")
 
-   end function d2d_listing_get_unit
+   end function d2d_log_get_unit
 
    !
    ! Close the IO unit for the log if needed
    !
-   module subroutine d2d_listing_close_unit(io_unit)
+   module subroutine d2d_log_close_unit(io_unit)
 
       use iso_fortran_env, only: output_unit
 
@@ -88,12 +110,12 @@ contains
          if (ierror /= 0) call decomp_2d_abort(ierror, "Could not close log file")
       end if
 
-   end subroutine d2d_listing_close_unit
+   end subroutine d2d_log_close_unit
 
    !
    ! Print some information about decomp_2d
    !
-   module subroutine d2d_listing(given_io_unit)
+   module subroutine d2d_log(given_io_unit)
 
[…]
