+++ src/decomp_2d_init_fin.f90
@@ -148,11 +148,11 @@
      !
      ! Get the IO unit for decomp_2d setup
      !
-     iounit = d2d_listing_get_unit()
+     iounit = d2d_log_get_unit()
      !
      ! Print the decomp_2d setup
      !
-     call d2d_listing(iounit)
+     call d2d_log(iounit)
 
 #ifdef PROFILER
      ! Stop the timer for decomp_2d_init
