### 2023-11-13 #5813 @rfj82982
@diehlpk I think I made all the changes that @p-costa  required. I think he has also acknowledge this in this post. https://github.com/openjournals/joss-reviews/issues/5813#issuecomment-1802204149
