### 2023-11-13 #5813 @diehlpk
> @diehlpk does the new release need to include also the paper? We have intentionally kept the paper in a dedicated branch which we keep in sync with the main. Last point: the title of the paper has changed from the one at the top of this issue (as requested by reviewers). Does this create any issue?

No, the paper does not needs to be included.

Yes, the title can be updated.
