### 2023-11-14 e43e0ed
Rename define_revcbuf -> perform_transpose
