### 2023-11-14 4f25e50
Use constants in transpose submodules
NVIDIA (apparently) requires this as does AOCC. Cray (15) and
GNU (11.2) both don't raise errors by its inclusion
