+++ src/transpose_y_to_z.f90
@@ -3,10 +3,7 @@
 ! This file contains the routines that transpose data from Y to Z pencil
 submodule(decomp_2d) d2d_transpose_y_to_z
 
-#ifdef NVHPC
-   ! FIXME : Nvidia compiler does not compile the submodule without this
    use decomp_2d_constants, only: mytype
-#endif
 
    implicit none
 
