### 2023-11-14 129521f
Tidy transpose_y_to_x
Following strategy employed by transpose_z_to_y tidy the real
transpose implementation
