+++ src/transpose_y_to_x.f90
@@ -9,6 +9,7 @@ submodule(decomp_2d) d2d_transpose_y_to_x
 
 contains
 
+   ! Short transpose interface for real numbers with implicit decomposition.
    module subroutine transpose_y_to_x_real_short(src, dst)
 
       implicit none
@@ -20,6 +21,7 @@ contains
 
    end subroutine transpose_y_to_x_real_short
 
+   ! Full transpose interface for real numbers with explicit decomposition
    module subroutine transpose_y_to_x_real_long(src, dst, decomp)
 
       implicit none
@@ -28,30 +30,29 @@ contains
       real(mytype), dimension(:, :, :), intent(OUT) :: dst
       TYPE(DECOMP_INFO), intent(IN) :: decomp
 
-#if defined(_GPU)
-      integer :: istat, nsize
+      real(mytype), dimension(:), pointer, contiguous :: wk1
+      real(mytype), dimension(:), pointer, contiguous :: wk2
+#ifdef _GPU
+      attributes(device) :: wk1, wk2
 #endif
+      
 
 #ifdef PROFILER
       if (decomp_profiler_transpose) call decomp_profiler_start("transp_y_x_r")
 #endif
 
+#ifdef _GPU
+      wk1 => work1_r_d
+      wk2 => work2_r_d
+#else
+      wk1 => work1_r
+      wk2 => work2_r
+#endif
+
       if (dims(1) == 1) then
-#if defined(_GPU)
-         nsize = product(decomp%ysz)
-         !$acc host_data use_device(src,dst)
-         istat = cudaMemcpy(dst, src, nsize, cudaMemcpyDeviceToDevice)
-         !$acc end host_data
-         if (istat /= 0) call decomp_2d_abort(__FILE__, __LINE__, istat, "cudaMemcpy")
-#else
-         dst = src
-#endif
+         call transpose_y_to_x_real_fast(src, dst, decomp)
       else
-#if defined(_GPU)
-         call transpose_y_to_x_real(src, dst, decomp, work1_r_d, work2_r_d)
-#else
-         call transpose_y_to_x_real(src, dst, decomp, work1_r, work2_r)
-#endif
+         call transpose_y_to_x_real(src, dst, decomp, wk1, wk2)
       end if
 
 #ifdef PROFILER
@@ -60,7 +61,34 @@ contains
 
    end subroutine transpose_y_to_x_real_long
 
-   subroutine transpose_y_to_x_real(src, dst, decomp, wk1, wk2)
+   ! Fast implementation of transpose for real numbers avoiding communication
+   subroutine transpose_y_to_x_real_fast(src, dst, decomp)
+
+      implicit none
+
+      real(mytype), dimension(:, :, :), intent(IN) :: src
+      real(mytype), dimension(:, :, :), intent(OUT) :: dst
+      TYPE(DECOMP_INFO), intent(IN) :: decomp
+
+      integer :: nsize
+#ifdef _GPU
+      integer :: istat
+#endif
+      
+      nsize = product(decomp%ysz)
+#if defined(_GPU)
+      !$acc host_data use_device(src,dst)
+      istat = cudaMemcpy(dst, src, nsize, cudaMemcpyDeviceToDevice)
+      !$acc end host_data
+      if (istat /= 0) call decomp_2d_abort(__FILE__, __LINE__, istat, "cudaMemcpy")
+#else
+      dst = src
+#endif
+
+    end subroutine transpose_y_to_x_real_fast
+
+    ! Default implementation of transpose for real numbers
+    subroutine transpose_y_to_x_real(src, dst, decomp, wk1, wk2)
 
       implicit none
 
@@ -73,7 +101,6 @@ contains
 #endif
 
       integer :: s1, s2, s3, d1, d2, d3
-      integer :: ierror
 
[…]
