### 2023-11-14 1137bf8
Remove unused preprocessor variable
