### 2023-11-14 cc6a06a
Remove unused variables
