### 2023-11-14 7638079
Extract a common implementation of transpose_real_fast
