### 2023-11-16 931842d
Avoid nested preprocessor macros in transpose subroutines
