+++ src/transpose_z_to_y.f90
@@ -115,10 +115,8 @@ contains
                            DECOMP_2D_COMM_ROW, ierror)
       end if
       if (ierror /= 0) call decomp_2d_abort(__FILE__, __LINE__, ierror, "MPI_ALLTOALL")
-#else
 
-#if defined(_GPU)
-#if defined(_NCCL)
+#elif defined(_NCCL)
       call decomp_2d_nccl_send_recv_row(wk2, &
                                         wk1, &
                                         decomp%z2disp, &
@@ -126,12 +124,13 @@ contains
                                         decomp%y2disp, &
                                         decomp%y2cnts, &
                                         dims(2))
-#else
+
+#elif defined(_GPU)
       call MPI_ALLTOALLV(wk1, decomp%z2cnts, decomp%z2disp, real_type, &
                          wk2, decomp%y2cnts, decomp%y2disp, real_type, &
                          DECOMP_2D_COMM_ROW, ierror)
       if (ierror /= 0) call decomp_2d_abort(__FILE__, __LINE__, ierror, "MPI_ALLTOALLV")
-#endif
+
 #else
       associate (wk => wk1)
       end associate
@@ -139,7 +138,6 @@ contains
                          wk2, decomp%y2cnts, decomp%y2disp, real_type, &
                          DECOMP_2D_COMM_ROW, ierror)
       if (ierror /= 0) call decomp_2d_abort(__FILE__, __LINE__, ierror, "MPI_ALLTOALLV")
-#endif
 #endif
 
       ! rearrange receive buffer
@@ -253,10 +251,8 @@ contains
                            DECOMP_2D_COMM_ROW, ierror)
       end if
       if (ierror /= 0) call decomp_2d_abort(__FILE__, __LINE__, ierror, "MPI_ALLTOALL")
-#else
 
-#if defined(_GPU)
-#if defined(_NCCL)
+#elif defined(_NCCL)
       call decomp_2d_nccl_send_recv_row(wk2, &
                                         wk1, &
                                         decomp%z2disp, &
@@ -265,12 +261,13 @@ contains
                                         decomp%y2cnts, &
                                         dims(2), &
                                         decomp_buf_size)
-#else
+
+#elif defined(_GPU)
       call MPI_ALLTOALLV(wk1, decomp%z2cnts, decomp%z2disp, complex_type, &
                          wk2, decomp%y2cnts, decomp%y2disp, complex_type, &
                          DECOMP_2D_COMM_ROW, ierror)
       if (ierror /= 0) call decomp_2d_abort(__FILE__, __LINE__, ierror, "MPI_ALLTOALLV")
-#endif
+
 #else
       associate (wk => wk1)
       end associate
@@ -278,8 +275,6 @@ contains
                          wk2, decomp%y2cnts, decomp%y2disp, complex_type, &
                          DECOMP_2D_COMM_ROW, ierror)
       if (ierror /= 0) call decomp_2d_abort(__FILE__, __LINE__, ierror, "MPI_ALLTOALLV")
-#endif
-
 #endif
 
       ! rearrange receive buffer

[…]
