### 2023-11-20 #5813 @rfj82982
@diehlpk  Updated branches are

- paper https://github.com/2decomp-fft/2decomp-fft/tree/SubmissionJOSS
- main https://github.com/2decomp-fft/2decomp-fft/tree/main
- release https://github.com/2decomp-fft/2decomp-fft/releases/tag/v2.0.1
