### 2023-11-21 #5813 @danielskatz
@rfj82982 - I'm the track editor who will complete the processing of this submission. I've found a number of small issues in the proof, with suggested changes in https://github.com/2decomp-fft/2decomp-fft/pull/299. Please merge this, or let me know what you disagree with, then we can proceed to acceptance and publication.
