### 2023-11-21 #5813 @danielskatz
Congratulations to @rfj82982 (Stefano Rolfo) and co-authors on your publication!!

JOSS depends on volunteers and we couldn't do this without you.
